envDataArray Get an array of environmental data from presence-absence points.

envDataArray Get an array of environmental data from presence-absence points.

Usage

envDataArray(occ, envData)

Arguments

occ

Occurence data frame. Should contain longitude and latitude columns

envData

List of environmental variables time series stack.

Value

An array of M points times N time steps times P environmental variables

Examples

occ <- mus_virtualis[1:5, ]
bio1_ts <- terra::unwrap(cmcc_cm_bio1)
envData <- list(bio1 = bio1_ts)
envDataArray(occ, envData)
#>         bio01   bio01.1   bio01.2    bio01.3   bio01.4   bio01.5   bio01.6
#> [1,] 2.041544 1.2718751 1.5751817  0.5713276 1.5104081 1.5082459 1.2022410
#> [2,] 4.734873 3.9701821 4.2729546  3.2682683 4.2139425 4.2010183 3.9038662
#> [3,] 3.321987 2.5879431 2.8549146  1.8579131 2.8482300 2.7758704 2.5128496
#> [4,] 1.232938 0.4855629 0.7518632 -0.2382688 0.7312467 0.6912795 0.4002289
#> [5,] 2.675358 1.9005995 2.2047160  1.2012200 2.1337284 2.1423210 1.8276727
#>        bio01.7     bio01.8   bio01.9    bio01.10   bio01.11   bio01.12 bio01.13
#> [1,] 1.3444174  0.71737323 1.6226010  0.71709225  0.6023474  0.5974347 1.798676
#> [2,] 4.0406282  3.42067502 4.3124088  3.41394589  3.3075438  3.2973179 4.490965
#> [3,] 2.6167950  2.05324462 2.8842523  2.02803843  1.9570990  1.8958019 3.065305
#> [4,] 0.5208363 -0.06377887 0.8072285 -0.06998784 -0.1654946 -0.2122951 0.979832
#> [5,] 1.9752697  1.34134326 2.2598258  1.34701134  1.2241749  1.2249785 2.433444
#>        bio01.14  bio01.15  bio01.16  bio01.17 bio01.18  bio01.19 bio01.20
#> [1,]  0.7477497 1.5523660 1.1831236 1.2242181 1.842464 1.4540200 1.994852
#> [2,]  3.4462284 4.2464069 3.8833775 3.9237369 4.539315 4.1519933 4.693018
#> [3,]  2.0596343 2.8401404 2.4800737 2.5286767 3.158748 2.7282700 3.296458
#> [4,] -0.0488958 0.7478101 0.3687964 0.4227769 1.061132 0.6278824 1.198548
#> [5,]  1.3769955 2.1857024 1.8103145 1.8519902 2.472404 2.0831894 2.623004
#>      bio01.21 bio01.22 bio01.23  bio01.24  bio01.25 bio01.26 bio01.27 bio01.28
#> [1,] 1.526184 1.981605 2.202857 1.3772860 0.9900415 1.666015 1.395033 2.720851
#> [2,] 4.224597 4.675181 4.902406 4.0772175 3.6940451 4.364479 4.090788 5.415594
#> [3,] 2.812925 3.231976 3.515872 2.6921556 2.3237131 2.980252 2.664208 4.013287
#> [4,] 0.710320 1.145488 1.411001 0.5824125 0.2052700 0.874549 0.566989 1.920695
#> [5,] 2.154628 2.614823 2.830099 2.0045132 1.6127578 2.294813 2.026398 3.353134
#>      bio01.29
#> [1,] 2.319535
#> [2,] 5.016120
#> [3,] 3.623733
#> [4,] 1.527008
#> [5,] 2.950044