Long-term stochastic growth rate worker function for the xsdm model

Computes the negative of the long-term stochastic growth rate, plus log(lambda_max), for the xsdm model, for each location. This is the fast version of a worker function that relies on C++ code, optimized using RcppParallel. The legacy pure-R reference is preserved as xsdm:::like_neg_ltsgr_r for testing and comparison.

Usage

like_neg_ltsgr(
  env_dat,
  mu,
  sigltil,
  sigrtil,
  o_mat,
  num_threads = RcppParallel::defaultNumThreads()
)

Arguments

env_dat

The environmental data array, dimensions (number of locations) x (time series length) x (number of environmental variables). Must not contain missing values.

mu

Vector of optimal environmental values. Length p=dim(env_dat)[3]. Unconstrained real numbers.

sigltil

Vector specifying width of the growth-environment function. Length p=dim(env_dat)[3]. Positive real numbers, Inf entries also allowed.

sigrtil

Vector specifying width of the growth-environment function. Length p=dim(env_dat)[3]. Positive real numbers, Inf entries also allowed.

o_mat

An orthogonal matrix, dimensions p by p.

num_threads

Number of threads for parallel computation. Defaults to RcppParallel::defaultNumThreads().

Value

A vector of length equal to the number of locations, as described above.

Details

Internally, this function:

Reshapes the environmental data into a matrix.

Computes inverse (diagonal) matrices for asymmetry adjustments.

Calls the C++ function like_ltsg() for efficient likelihood computation.

Note

Ensure that env_dat has no missing values. The parameter vectors mu, sigltil, and sigrtil must have length equal to the number of environmental variables (p). The argument o_mat must be a p x p orthogonal matrix (i.e., o_mat %*% t(o_mat) is the identity).

Examples

# Example usage:
like_neg_ltsgr(env_dat = example_1$env_array,
               mu      = example_1$true_par_list$mu,
               sigltil = example_1$true_par_list$sigltil,
               sigrtil = example_1$true_par_list$sigrtil,
               o_mat   = example_1$true_par_list$o_mat)
#>    [1] 2.738171 8.535266 8.374321 7.154807 8.444485 3.477033 3.272479 8.034020
#>    [9] 5.494181 3.406080 3.894184 8.221686 7.075254 5.269100 4.288629 3.740807
#>   [17] 6.999191 4.930312 6.834511 2.327886 2.074068 4.776596 2.464446 6.709562
#>   [25] 8.304522 3.820160 8.435004 5.298979 6.876833 4.155158 6.926461 6.161269
#>   [33] 4.097653 5.102134 4.291076 6.345642 4.245375 3.341317 8.660052 6.664898
#>   [41] 2.810356 5.362398 6.999660 7.627527 4.651365 5.483883 4.516378 5.404858
#>   [49] 3.790042 6.025831 8.251108 6.783180 7.984423 4.287914 3.056667 6.856656
#>   [57] 8.214739 7.498210 5.261167 5.338618 8.011760 5.193644 4.798711 8.416415
#>   [65] 7.293172 3.600957 4.264491 8.319323 6.217450 4.367401 5.124876 1.979788
#>   [73] 7.917596 6.576324 3.495894 5.479428 7.474645 6.975485 3.053595 3.440059
#>   [81] 4.145340 4.927784 3.322202 4.472423 6.570589 7.014075 5.538776 2.680407
#>   [89] 4.693739 6.103064 6.448714 2.783318 5.408007 8.305771 3.643784 6.221024
#>   [97] 6.240697 7.991659 5.309896 3.645075 4.120219 8.335431 5.147934 3.522013
#>  [105] 8.410355 4.912260 4.622131 5.170946 7.813666 7.410260 7.780804 7.690245
#>  [113] 4.278769 4.433976 6.439727 4.491546 5.024642 2.591280 7.766783 5.951452
#>  [121] 6.646883 7.701346 3.647620 7.711426 4.535503 6.205746 4.217129 6.303538
#>  [129] 4.647464 8.543404 7.927106 7.634204 6.767605 5.450163 5.276290 8.677979
#>  [137] 4.104554 6.750432 5.101582 3.535671 4.327127 3.899442 2.513490 6.251582
#>  [145] 4.136154 8.488412 7.446858 2.584259 6.671945 3.851036 7.779567 8.397149
#>  [153] 3.530302 5.730067 7.852212 4.845694 4.944500 5.696837 8.223739 4.374859
#>  [161] 5.328191 7.272478 3.679817 6.775375 8.134278 5.630475 7.462363 8.346814
#>  [169] 5.211055 5.127361 4.745629 7.709655 6.691052 2.921612 6.464432 8.342831
#>  [177] 3.845642 8.447241 6.873000 5.316705 6.203020 2.769259 7.306254 2.951352
#>  [185] 8.288232 3.652701 6.782896 7.864908 6.941641 7.954834 7.279754 7.033723
#>  [193] 3.971800 3.557843 3.819581 5.665545 8.269367 8.213562 4.699454 5.075297
#>  [201] 7.673952 2.967565 5.479933 3.740539 7.866173 5.794756 3.110880 6.288045
#>  [209] 4.136789 5.034436 3.201250 3.983400 7.933329 6.077172 7.230290 6.730007
#>  [217] 5.377706 5.802934 3.670574 2.933928 5.957608 7.205059 3.711949 6.983483
#>  [225] 3.239194 6.335887 3.289620 6.014989 4.945292 3.322412 3.503462 7.155621
#>  [233] 5.889508 3.292937 8.294997 2.559470 3.128033 3.970428 5.658968 5.905711
#>  [241] 2.615392 2.456258 3.667483 5.087955 7.433216 7.586851 7.706441 5.496604
#>  [249] 6.784210 4.197012 3.969127 7.780620 4.569003 3.739359 7.585584 6.283181
#>  [257] 2.361769 7.290611 6.927343 7.520906 8.527811 6.068080 7.204822 4.434878
#>  [265] 2.492797 8.022582 8.027586 4.773539 6.594982 2.860595 7.013509 8.685222
#>  [273] 6.542163 2.561165 2.586524 6.109590 2.760930 8.026780 3.050958 1.844593
#>  [281] 8.163568 3.374663 6.314393 7.206267 5.450465 6.726003 7.282657 7.573248
#>  [289] 4.536227 7.985194 4.070292 4.351968 5.404378 5.499150 5.836856 6.364855
#>  [297] 3.162164 5.752320 6.161986 7.273024 5.235339 8.622596 2.973380 3.124899
#>  [305] 2.943223 6.379983 6.403187 7.425073 8.422545 5.931950 6.899647 6.816617
#>  [313] 6.971037 3.377473 4.321292 3.693707 4.793186 3.900874 2.681686 8.605958
#>  [321] 4.206489 7.616487 7.673731 7.714181 6.153605 4.510407 2.689889 6.279882
#>  [329] 4.343616 6.426235 6.087054 3.416215 7.227908 4.843000 2.985486 2.792633
#>  [337] 3.307358 7.258635 7.551705 3.969658 4.025248 5.751400 2.035704 6.497565
#>  [345] 7.933129 5.234216 5.687010 2.539760 5.279772 5.829322 6.294680 3.272314
#>  [353] 7.152536 6.923463 6.968193 6.619868 3.948832 8.679205 6.135075 5.950581
#>  [361] 3.456851 4.783452 3.620395 4.837772 6.228984 3.439286 6.268120 3.927663
#>  [369] 2.447301 7.229603 7.322477 3.821366 7.705315 8.036480 5.540306 7.080314
#>  [377] 5.447749 5.095901 2.527407 6.584077 3.508366 4.682486 4.101918 3.799091
#>  [385] 4.973119 3.985040 3.729906 5.554829 7.094541 8.019271 2.329663 6.167402
#>  [393] 5.599354 3.247175 5.518319 8.286294 8.671559 8.665080 3.020188 4.232752
#>  [401] 3.451821 3.451614 8.401605 2.717455 7.915186 6.504449 2.680351 7.584430
#>  [409] 8.101095 7.715503 7.358174 3.877302 3.030461 2.558614 6.942178 4.847294
#>  [417] 4.189646 3.237823 5.282576 3.985766 2.112286 7.474094 3.878193 5.340085
#>  [425] 6.494366 4.547283 5.160111 6.072133 2.160651 3.737604 4.948909 1.936117
#>  [433] 7.845225 2.615043 5.006440 6.472571 8.186124 4.570254 8.066034 3.448542
#>  [441] 7.934359 7.895704 6.997027 7.890756 6.198348 3.660974 4.823904 7.650551
#>  [449] 7.833773 4.227792 6.156741 4.733819 8.015880 4.603286 3.545889 6.032412
#>  [457] 4.575272 6.251188 7.822554 4.084764 5.621771 3.505276 4.075021 5.244029
#>  [465] 7.060772 5.471940 6.861719 4.819562 2.956816 7.756185 8.016243 2.872487
#>  [473] 4.123046 2.986877 6.151864 5.009757 7.639867 3.309031 3.169437 6.796955
#>  [481] 5.730892 3.721554 8.480516 6.774152 4.392391 3.517795 6.480300 6.207808
#>  [489] 4.784948 2.357236 4.293837 5.003840 8.260322 6.167962 6.685890 5.986726
#>  [497] 4.439560 5.268185 3.692445 4.995306 8.440351 6.064291 7.957908 5.430812
#>  [505] 7.770940 8.281199 3.646238 6.410633 7.655068 5.583325 7.590048 5.729619
#>  [513] 3.557723 5.335095 3.581894 3.319113 6.545221 7.913582 8.127177 5.752468
#>  [521] 7.685988 5.670282 4.941975 5.848321 3.864835 3.503714 3.982686 7.273977
#>  [529] 3.939822 4.406699 5.671054 7.422249 3.043856 5.786128 2.259710 5.274936
#>  [537] 5.710916 5.117274 5.661057 4.056405 4.590258 7.912187 4.334486 7.912549
#>  [545] 2.525762 8.063547 4.318433 4.853824 4.865807 7.624448 5.712030 5.489711
#>  [553] 3.967668 4.848766 6.506274 5.443358 8.577216 7.635640 5.873814 6.744487
#>  [561] 4.973482 2.775246 8.286285 7.922041 8.471732 4.307777 8.196540 6.993114
#>  [569] 4.157498 7.781853 7.558540 3.075448 4.393681 7.980965 4.469846 4.658037
#>  [577] 5.480192 3.430931 4.090203 3.912564 2.638866 3.212021 5.162167 5.341303
#>  [585] 5.499401 5.847025 4.660721 3.823180 5.767246 7.377622 2.930334 3.374140
#>  [593] 5.982484 2.714523 7.547959 7.244319 5.609992 4.784291 6.344278 7.084156
#>  [601] 4.802776 4.274909 6.449187 3.427608 3.693117 4.069509 5.456143 3.669372
#>  [609] 5.619555 5.913603 8.476933 3.498282 5.432434 8.408024 8.596682 3.182762
#>  [617] 4.300191 6.159364 4.843586 4.841264 5.370967 3.821050 4.362531 7.689048
#>  [625] 6.665243 7.154496 4.033387 3.555588 4.593265 3.377781 6.919569 7.298779
#>  [633] 5.422203 4.555155 6.737008 4.172477 2.553697 6.720496 4.164538 2.596982
#>  [641] 8.347730 2.785452 8.259568 4.302055 2.352061 4.563356 4.646854 6.553449
#>  [649] 7.443764 7.206827 6.304581 4.618992 3.244508 8.018727 4.766374 4.769661
#>  [657] 5.961189 4.817558 7.126731 4.181735 6.490961 5.532172 7.067334 3.295478
#>  [665] 7.676368 2.496864 5.300190 4.988234 7.448746 7.893728 6.219486 6.874343
#>  [673] 5.505493 7.168542 8.511330 3.726587 3.764189 5.913508 4.939287 3.778838
#>  [681] 4.851821 8.122428 2.916222 6.696599 6.428615 4.561121 5.282771 6.382401
#>  [689] 4.725508 6.023216 4.346883 3.913192 5.304983 6.179082 6.734587 4.982478
#>  [697] 5.648537 5.677169 3.188524 8.595754 5.435996 7.867928 6.716072 8.012291
#>  [705] 6.670052 4.191274 5.503510 4.450612 6.247518 3.287386 3.797807 2.658075
#>  [713] 3.285803 6.342421 5.439417 6.993322 6.756449 4.612762 3.195821 5.302823
#>  [721] 5.575945 5.568095 7.792703 6.585643 7.483930 4.074094 5.666728 7.088191
#>  [729] 3.122774 6.049405 2.409910 4.090649 4.903148 7.570804 8.474624 4.770945
#>  [737] 4.154927 5.334148 6.734582 4.504836 4.523618 3.278765 5.442380 3.066054
#>  [745] 5.866659 6.170346 3.808898 6.223310 7.027555 8.021163 7.068678 4.851655
#>  [753] 7.773158 7.231622 8.090406 3.312508 8.166651 8.480308 3.025115 5.567738
#>  [761] 7.806375 8.507377 4.539791 3.231424 2.704578 7.825126 6.217286 6.713378
#>  [769] 5.634191 6.494820 3.397384 4.868878 3.726779 5.389809 4.076884 4.277643
#>  [777] 3.672147 3.998512 8.021931 7.838774 3.904019 5.091911 6.722692 6.083745
#>  [785] 5.248950 8.037602 3.238276 5.888718 4.236371 5.590665 8.668876 3.990844
#>  [793] 4.194881 8.379268 3.494161 3.836587 7.120971 7.651745 5.233000 7.799905
#>  [801] 4.267790 4.726320 7.565624 2.667601 4.301659 5.626010 3.534682 7.736772
#>  [809] 2.358384 6.294209 2.887388 4.085603 2.767285 3.562569 2.754517 4.961691
#>  [817] 5.095786 6.523526 8.224669 6.337783 8.053719 4.533643 3.175342 2.494171
#>  [825] 5.932482 8.085847 4.204271 5.116782 2.709875 5.831805 2.735037 7.964228
#>  [833] 6.569012 8.115789 3.135490 8.110443 7.832184 3.689198 7.191854 3.034071
#>  [841] 2.842213 7.979132 3.830341 5.214407 3.381384 8.616595 5.219445 8.271040
#>  [849] 3.862106 7.888417 6.245119 4.562539 3.699590 2.962022 7.240975 7.161624
#>  [857] 5.060736 7.313176 3.288401 5.885004 4.564450 7.303013 6.814671 8.475478
#>  [865] 7.860754 8.451341 4.688513 4.678283 5.727080 5.170603 4.291227 6.475604
#>  [873] 2.603295 2.760989 7.459558 8.102027 7.461878 2.410126 7.807386 7.194237
#>  [881] 6.086081 3.552605 7.114467 7.744252 8.519727 8.272525 5.014570 4.142912
#>  [889] 4.224112 6.343662 2.820095 2.939107 5.143522 8.191626 7.272415 4.223209
#>  [897] 7.604237 7.504105 4.957457 2.690517 4.707225 5.840925 6.673861 7.387350
#>  [905] 3.126374 6.841601 8.469149 5.023942 6.834962 8.106575 6.921475 3.870649
#>  [913] 5.548968 7.394421 7.157734 3.242620 4.951736 4.804114 6.130948 6.544578
#>  [921] 4.533288 5.959637 8.362821 4.600947 6.755693 2.414515 4.478275 3.900118
#>  [929] 6.463981 6.301487 5.571437 8.191842 8.676462 8.067420 2.158200 5.778746
#>  [937] 5.974072 6.132136 8.461567 3.152480 7.282938 6.736813 6.954803 5.222599
#>  [945] 8.590609 8.580312 6.749103 4.947143 5.073389 4.508214 8.585070 3.383786
#>  [953] 7.074032 4.921445 4.155504 5.249568 6.704727 4.163071 4.054647 5.207599
#>  [961] 5.223112 6.998289 3.515707 6.064597 7.702133 8.125316 8.536342 6.053649
#>  [969] 3.477018 6.206682 6.722753 2.224025 8.178045 4.694019 8.346858 6.135494
#>  [977] 3.704659 5.976697 5.085363 3.403459 8.200203 5.356457 6.778927 4.348040
#>  [985] 6.057429 4.620366 7.120062 3.405037 2.839946 2.746828 2.727760 3.774420
#>  [993] 3.470485 3.305958 4.103848 5.966980 7.354790 6.200861 4.372938 7.222016
#>  [ reached 'max' / getOption("max.print") -- omitted 3000 entries ]