envDataArray Get an array of environmental data from presence-absence points.

envDataArray Get an array of environmental data from presence-absence points.

Usage

envDataArray(envData, occ = NULL)

Arguments

envData

List of environmental variables time series stack.

occ

Occurence data frame. Should contain longitude and latitude columns

Value

An array of M points times N time steps times P environmental variables

Examples

occ <- mus_virtualis[1:5, ]
bio1_ts <- terra::unwrap(cmcc_cm_bio1)
bio12_ts <- terra::unwrap(cmcc_cm_bio12)
envData <- list(bio1 = bio1_ts, bio12 = bio12_ts)
envDataArray(envData, occ)
#> , , bio1
#> 
#>       bio01_1   bio01_2   bio01_3    bio01_4   bio01_5   bio01_6   bio01_7
#> [1,] 2.041544 1.2718751 1.5751817  0.5713276 1.5104081 1.5082459 1.2022410
#> [2,] 4.734873 3.9701821 4.2729546  3.2682683 4.2139425 4.2010183 3.9038662
#> [3,] 3.321987 2.5879431 2.8549146  1.8579131 2.8482300 2.7758704 2.5128496
#> [4,] 1.232938 0.4855629 0.7518632 -0.2382688 0.7312467 0.6912795 0.4002289
#> [5,] 2.675358 1.9005995 2.2047160  1.2012200 2.1337284 2.1423210 1.8276727
#>        bio01_8     bio01_9  bio01_10    bio01_11   bio01_12   bio01_13 bio01_14
#> [1,] 1.3444174  0.71737323 1.6226010  0.71709225  0.6023474  0.5974347 1.798676
#> [2,] 4.0406282  3.42067502 4.3124088  3.41394589  3.3075438  3.2973179 4.490965
#> [3,] 2.6167950  2.05324462 2.8842523  2.02803843  1.9570990  1.8958019 3.065305
#> [4,] 0.5208363 -0.06377887 0.8072285 -0.06998784 -0.1654946 -0.2122951 0.979832
#> [5,] 1.9752697  1.34134326 2.2598258  1.34701134  1.2241749  1.2249785 2.433444
#>        bio01_15  bio01_16  bio01_17  bio01_18 bio01_19  bio01_20 bio01_21
#> [1,]  0.7477497 1.5523660 1.1831236 1.2242181 1.842464 1.4540200 1.994852
#> [2,]  3.4462284 4.2464069 3.8833775 3.9237369 4.539315 4.1519933 4.693018
#> [3,]  2.0596343 2.8401404 2.4800737 2.5286767 3.158748 2.7282700 3.296458
#> [4,] -0.0488958 0.7478101 0.3687964 0.4227769 1.061132 0.6278824 1.198548
#> [5,]  1.3769955 2.1857024 1.8103145 1.8519902 2.472404 2.0831894 2.623004
#>      bio01_22 bio01_23 bio01_24  bio01_25  bio01_26 bio01_27 bio01_28 bio01_29
#> [1,] 1.526184 1.981605 2.202857 1.3772860 0.9900415 1.666015 1.395033 2.720851
#> [2,] 4.224597 4.675181 4.902406 4.0772175 3.6940451 4.364479 4.090788 5.415594
#> [3,] 2.812925 3.231976 3.515872 2.6921556 2.3237131 2.980252 2.664208 4.013287
#> [4,] 0.710320 1.145488 1.411001 0.5824125 0.2052700 0.874549 0.566989 1.920695
#> [5,] 2.154628 2.614823 2.830099 2.0045132 1.6127578 2.294813 2.026398 3.353134
#>      bio01_30
#> [1,] 2.319535
#> [2,] 5.016120
#> [3,] 3.623733
#> [4,] 1.527008
#> [5,] 2.950044
#> 
#> , , bio12
#> 
#>       bio01_1  bio01_2  bio01_3  bio01_4  bio01_5  bio01_6  bio01_7  bio01_8
#> [1,] 338.0783 491.8341 469.4761 478.5737 481.8249 329.6221 492.0890 452.8594
#> [2,] 313.3871 448.4263 425.1769 435.2373 436.7535 298.1498 448.8610 411.6198
#> [3,] 296.2627 402.9470 381.8153 400.5392 392.4493 265.6382 411.2113 367.3802
#> [4,] 325.0968 461.2005 439.8220 448.9147 446.5277 301.9816 464.4093 418.4908
#> [5,] 347.0207 511.4862 490.1235 494.3226 504.4770 342.4032 510.7664 473.5115
#>       bio01_9 bio01_10 bio01_11 bio01_12 bio01_13 bio01_14 bio01_15 bio01_16
#> [1,] 676.8963 449.3065 505.1880 619.0507 525.7949 564.1982 757.7330 417.1659
#> [2,] 618.0046 409.9378 458.0567 559.1728 478.3410 512.1521 690.8510 378.7396
#> [3,] 569.6613 373.3756 413.8610 488.3848 428.9124 468.7120 621.6374 350.0553
#> [4,] 639.4309 421.6475 471.7987 571.8433 485.1405 534.7005 711.2036 395.4493
#> [5,] 699.4839 467.0876 526.8031 648.4700 545.5829 586.8502 788.2558 432.9401
#>      bio01_17 bio01_18 bio01_19 bio01_20 bio01_21 bio01_22 bio01_23 bio01_24
#> [1,] 507.5783 394.1751 433.8776 373.9539 644.3779 439.8986 417.8187 473.7581
#> [2,] 462.3387 355.8710 394.5806 337.6175 586.3249 401.5230 380.5562 428.7120
#> [3,] 416.0023 310.7604 362.1224 298.2097 533.9747 368.8502 348.1691 388.7051
#> [4,] 475.0484 360.7419 410.2258 342.4885 607.7765 417.1682 392.2681 447.3963
#> [5,] 528.1590 415.7627 449.5895 391.6705 673.0046 454.6797 435.4983 495.3134
#>      bio01_25 bio01_26 bio01_27 bio01_28 bio01_29 bio01_30
#> [1,] 477.4332 484.2604 521.8420 407.7120 498.8041 452.3180
#> [2,] 436.0323 440.9816 473.7086 366.4401 455.5968 411.9424
#> [3,] 391.6313 398.3341 434.3192 329.0069 413.2857 375.2627
#> [4,] 443.9032 452.5553 492.1924 376.1820 468.3065 423.6129
#> [5,] 498.1498 503.9770 541.4550 420.5576 517.5599 466.0991
#>