Long-term stochastic growth rate worker function for the xsdm model

Computes the negative of the long-term stochastic growth rate, plus log(lambda_max), for the xsdm model, for each location. This is the fast version of a worker function that relies on C++ code, optimized using RcppParallel. The legacy pure-R reference is preserved as xsdm:::like_neg_ltsgr_r for testing and comparison.

Usage

like_neg_ltsgr(
  env_dat,
  mu,
  sigltil,
  sigrtil,
  o_mat,
  num_threads = RcppParallel::defaultNumThreads()
)

Arguments

env_dat

The environmental data array, dimensions (number of locations) x (time series length) x (number of environmental variables). Must not contain missing values.

mu

Vector of optimal environmental values. Length p=dim(env_dat)[3]. Unconstrained real numbers.

sigltil

Vector specifying width of the growth-environment function. Length p=dim(env_dat)[3]. Positive real numbers, Inf entries also allowed.

sigrtil

Vector specifying width of the growth-environment function. Length p=dim(env_dat)[3]. Positive real numbers, Inf entries also allowed.

o_mat

An orthogonal matrix, dimensions p by p.

num_threads

Number of threads for parallel computation. Defaults to RcppParallel::defaultNumThreads().

Value

A vector of length equal to the number of locations, as described above.

Details

Internally, this function:

Reshapes the environmental data into a matrix.

Computes inverse (diagonal) matrices for asymmetry adjustments.

Calls the C++ function like_ltsg() for efficient likelihood computation.

Note

Ensure that env_dat has no missing values. The parameter vectors mu, sigltil, and sigrtil must have length equal to the number of environmental variables (p). The argument o_mat must be a p x p orthogonal matrix (i.e., o_mat %*% t(o_mat) is the identity).

Examples

# Example usage:
like_neg_ltsgr(env_dat = example_1$env_array,
               mu      = example_1$par_list$mu,
               sigltil = example_1$par_list$sigltil,
               sigrtil = example_1$par_list$sigrtil,
               o_mat   = example_1$par_list$o_mat)
#>    [1]   14.944779   16.840187   15.696630   13.752505   15.865081   14.311879
#>    [7]   15.506017   12.757522   13.256893   16.687851   16.142675   16.474053
#>   [13]   12.113523   16.364515   13.798161   16.232936   16.712399   16.742344
#>   [19]   14.471396   13.424058   12.459628   11.124633   11.917702   15.841068
#>   [25]   16.944757   16.554079   16.741006   13.562687   16.816864   16.093501
#>   [31]   14.779380   14.004655   15.569579   16.138915   14.581040   15.066121
#>   [37]   17.156586   15.212144   14.877289   16.085464   15.755465   12.958554
#>   [43]   15.635476   15.436973   15.154261   16.345117   12.149733   14.911328
#>   [49]   16.710546   15.133827   13.085495   15.757768   13.329035   14.496302
#>   [55]   13.948137   10.903962    8.870944   16.761451   12.040174   13.071297
#>   [61]   14.100997   13.348911   15.365888   16.428481   16.686901   16.811088
#>   [67]    9.403849   16.179156   14.985875   16.452945   16.706488   15.491487
#>   [73]   16.586517   16.917900   15.940605   11.529352   13.946782   13.860951
#>   [79]   15.820501   15.804868   13.834755   13.916973   15.507983   16.892498
#>   [85]   14.179690   15.461442   16.665603   15.019199   13.039092   13.935716
#>   [91]   17.345254   14.323201   14.389450    9.938209   16.060483   11.995816
#>   [97]   12.692997   15.965450   16.301970   13.661763   16.924619   12.514256
#>  [103]   17.278262   16.953321   16.207536   12.066094   15.962079   13.435920
#>  [109]   16.807542   16.227971   15.505003   12.866289   12.927957   16.961878
#>  [115]   13.307410   13.155922   15.729319   11.855164   13.282193   11.998660
#>  [121]   12.958098   15.887786   13.546298   14.454935   17.216729   13.812642
#>  [127]   13.439336   14.578823   15.530059   15.620056   16.514382   14.647392
#>  [133]   12.884229   17.068559   15.997399   14.673406   14.451937   16.257757
#>  [139]   12.495286   11.827606   16.968537   17.166713   12.571579   13.436431
#>  [145]   15.888743   12.272577   13.689791   16.315204   14.063139   14.335760
#>  [151]   16.250448   15.001531   13.243809   13.651519   13.086794   17.004533
#>  [157]   13.523820   15.227094   14.796082   11.618472   12.531688   14.268100
#>  [163]   14.108676   15.632411   15.304729   15.856212   16.139569   13.933093
#>  [169]   17.006286   12.483222   17.322272   16.799464   15.190861   17.265494
#>  [175]   16.818504   15.581760   15.819397   14.062071   16.988819   11.226471
#>  [181]   14.537267   16.388174   17.260698   15.049557   17.048653   13.719937
#>  [187]   17.156395   14.045266   16.412975   12.676821   16.717812   12.274450
#>  [193]   15.390613   16.606900   13.843460   12.291472   15.572324   11.777329
#>  [199]   13.333549   14.053582   16.032232   15.817078   14.732042   13.183180
#>  [205]   13.683098   12.115166   11.952681   15.097066   16.899637   11.973454
#>  [211]   16.470329   12.573078   14.298295   14.570562   16.439756   12.973748
#>  [217]   12.534619   10.857181   15.778485   16.264453   17.366156   12.788164
#>  [223]   15.844085   12.657126   16.487526   15.322961   15.797193   13.748549
#>  [229]   13.521046   15.159317   13.729902   13.860204   16.755871   15.362463
#>  [235]   16.732744   16.984848   17.254564   10.036624   17.238819   13.486905
#>  [241]   13.511684   15.981243   16.014920   14.022854   16.945155   14.748358
#>  [247]   14.130954   14.343465   16.459298   12.217344   14.130682   15.744441
#>  [253]   15.308331   14.361640   14.943351   16.743162   14.308288   16.220911
#>  [259]   17.111107   16.676498   15.710137   12.736604   14.215595   16.902963
#>  [265]   12.736853   15.485161   14.938385   12.925008   13.888123   15.536108
#>  [271]   16.660813   11.801373   16.199482   12.324159   12.444379   16.252511
#>  [277]    8.486360   14.326537   15.903981   14.318991   13.855571    8.848681
#>  [283]   16.420372   15.458468   16.926782   12.252662    9.454311   17.358476
#>  [289]   16.026212   13.867454   16.589046   15.815603   14.389222   12.354089
#>  [295]    9.452423   12.191239   16.558264   15.211960   12.937710   13.244358
#>  [301]   13.634272   14.878419   17.364821   16.660673   13.326660   15.940564
#>  [307]   11.927775   13.538675   10.019272   15.969404   12.436335   15.535654
#>  [313]   13.763180   14.056749   15.274375   14.599913   15.073332   14.068450
#>  [319]   10.712416   14.577456   17.306157   13.830143   16.958466   13.817897
#>  [325]   14.544077   12.933398   14.371002   14.644813   16.871927   16.942302
#>  [331]   12.310409   12.774137   13.917924   16.340074   16.677046   13.392016
#>  [337]   16.434532   16.137299   11.203695   16.637232   15.565606   14.419766
#>  [343]   14.757848   13.960203   11.561177   10.784738   11.330284   12.604005
#>  [349]   14.286970   13.956648   12.421572   14.714940   16.956385   14.612972
#>  [355]   15.106863   14.789864   16.772550   14.290476   15.505877   14.207152
#>  [361]   15.771884   16.068159   12.696035   15.534609   16.690879   13.555882
#>  [367]   14.872340   12.171733   13.364591   11.149713   14.601033   17.362928
#>  [373]   12.787275   15.521026   13.741743   15.707165   14.303494   13.235845
#>  [379]   11.909890   15.619972   12.553355   15.031631   14.831713   13.680348
#>  [385]   12.764483   12.318455   12.577978   13.881372   16.212875   13.097470
#>  [391]   15.512458   15.424389   14.579275   12.733235   14.831925   11.487007
#>  [397]   12.875777   13.594155   16.673920   15.581991   11.937892   13.359198
#>  [403]   14.623169   17.191717   16.968043   15.462882    9.400503   15.706891
#>  [409]   12.567324   14.290292   15.046025   17.061948   16.599460   13.204506
#>  [415]   13.727139   14.234638   17.352970   16.398193   15.345592   17.080894
#>  [421]   16.718842   13.702181   16.339569   14.843460   12.016830   10.998656
#>  [427]   12.222006    9.506809   13.861817   15.111523   17.194146   10.898403
#>  [433]   10.718997   14.802217   13.447009   16.541671   10.530250   12.300536
#>  [439]   15.862641   13.726902   13.525927   13.706391   15.346513   14.486600
#>  [445]   16.625285   15.770394   16.505957   13.475434   14.411798    8.697081
#>  [451]   17.370297   17.125192   15.775518   14.553986   15.834084   16.626327
#>  [457]   16.124484   15.544032   14.857598   13.268725   16.514747   15.301612
#>  [463]   13.158610   13.985368   13.650283   11.378490   16.469719   12.831279
#>  [469]   13.598876   16.996788   17.343584   12.878953   17.036219   13.514452
#>  [475]   16.911857   15.059023   15.820735   12.080047   13.984909   13.485643
#>  [481]   17.106967   15.478169   17.197782   15.266738   14.165024   13.950318
#>  [487]   17.301368   16.598720   14.467023   17.113028   15.370062   17.332839
#>  [493]   14.815032   10.822429   12.509900   16.626046   14.444809    9.792303
#>  [499]   16.727031   15.295454   17.191025   15.265814   16.848987   16.016245
#>  [505]   14.236524   17.031597   13.175616   14.440886   12.905769   13.385431
#>  [511]   17.116299   14.483900   15.255113   11.811596   16.938718   15.701579
#>  [517]   12.662419   15.934332   12.950749   16.300877   13.352350   14.293417
#>  [523]   17.213597   14.753283   12.524146   11.060059   13.224649   16.852959
#>  [529]   16.298286   13.852990   15.667211   14.492880   15.815677   17.077877
#>  [535]   14.883068   16.313823   17.251139   15.705694   15.168812   13.317222
#>  [541]   17.326792   16.590130   13.563182   16.679810   12.518989   12.231382
#>  [547]   16.947883   14.011506   11.524321   12.631239   12.568346   15.314215
#>  [553]   12.981833   15.264722   16.687774   13.891354   17.075787   16.630436
#>  [559]   17.292428   17.340022   14.948305   16.962649   17.126104   11.697597
#>  [565]   16.355233   13.640871   14.965536   14.072133   15.762534   13.600040
#>  [571]   16.918822   17.261080   14.852296   12.255417   15.274022   15.970573
#>  [577]   17.294612   13.850936   16.516048   15.930329   11.057825   16.884230
#>  [583]   14.456237   14.145254   13.457758   13.149753   17.169936   16.692886
#>  [589]   13.326508   12.134847   11.589934   10.084502   15.817693   16.091911
#>  [595]   11.587513   12.184400   16.534913   15.374795    9.881549   16.284521
#>  [601]   15.602547   14.145221   17.364038   15.795147   16.328572   15.031048
#>  [607]   17.306546   13.379693   14.498249   14.677216   16.530778   12.839082
#>  [613]   15.630084   14.132964   16.472986   16.662347   12.067454   13.550101
#>  [619]   15.023841   14.860556   16.939558   12.945957   15.754663   15.888552
#>  [625]   14.430664   14.446507   15.474152   15.498616   11.075064   17.230401
#>  [631]   12.622968   15.569222   13.161625   11.971449   14.197265   11.991887
#>  [637]   14.402523   12.615174   15.498039   15.692757   14.944369   14.900935
#>  [643]   14.726862   16.959932   15.778081   16.376596   14.927485   15.681680
#>  [649]   16.485204   14.782782   12.827989   14.422140   12.206159   15.095670
#>  [655]   14.304787   15.083244   14.348385   13.172301   14.572850   12.127974
#>  [661]   13.804398   16.984941   15.830212   12.384606   16.775627   16.540860
#>  [667]   14.918734   16.226320   16.438914   15.207019   11.835363   13.171425
#>  [673]   13.902563   14.722056   16.646448   14.273265   14.691967   14.774043
#>  [679]   12.456194   17.113750   13.274777   13.285094   12.703382   16.675759
#>  [685]   15.108316   16.748261   15.004865   12.457246   12.811630   13.658382
#>  [691]   14.349569   12.747727   12.711193   15.965951    9.603899   14.673196
#>  [697]   16.395026   13.662553   12.245703   17.326109   14.899110   15.061215
#>  [703]   10.161119   12.465910   15.368281   17.139033   13.483874   12.998057
#>  [709]   16.587677   17.115994   14.348303   15.018724   11.243194   11.631656
#>  [715]   16.393137   14.946442   15.336005   16.500318   13.764027   16.711206
#>  [721]   15.779928   14.801462   11.554225   15.856300   14.265408   12.619109
#>  [727]   10.723499   16.188719   16.649873   15.635967   17.321199   17.242579
#>  [733]   15.224562   15.273552   13.820146   14.649883   12.808055   16.913382
#>  [739]   16.341656   13.259485   17.219747   15.392442   13.833021   15.927048
#>  [745]   13.175017   13.620428   13.356012   17.358295   15.424067   17.255658
#>  [751]   14.341705   13.173112   10.104872   14.435403   14.479399   15.226674
#>  [757]   13.392845   15.426059   16.269883   14.598862   15.843420   13.886817
#>  [763]   16.258380   14.255430   15.262496   14.732117   17.055425   16.722233
#>  [769]   16.376241   15.913470   10.262614   13.276225   15.854156   15.278606
#>  [775]   13.335329   16.721647   14.866800   15.251602   12.365774   15.125548
#>  [781]   13.499867   16.135446    9.519780   11.169509   12.094893   15.625996
#>  [787]   11.795346   14.588686   17.073542   12.195455   17.311548   13.319714
#>  [793]   10.766837   15.014967   16.918802   14.225178   13.753903   14.920899
#>  [799]   14.677183   15.307088   16.532361   13.257383   13.349482   13.038392
#>  [805]   11.937246   12.471588   13.018200   12.269050   16.474313   15.040339
#>  [811]   17.258409   13.359335   13.550409   13.650788   17.376529   16.518553
#>  [817]   12.623086   13.407536   16.756264   16.256251   16.925688   12.957909
#>  [823]   16.431102   16.958262   16.191105   14.694791   16.645122   16.986318
#>  [829]   15.159239   15.823731   11.254426   15.371557   12.475111   15.051336
#>  [835]   17.361102   15.861700   13.486692   16.614935   14.214908   16.674070
#>  [841]   12.052644   13.161784   14.127122   17.159986   15.080599   11.683341
#>  [847]   16.837878   17.073535   17.061766   12.927235   13.681309   15.957821
#>  [853]   11.486428   15.571282   14.099756   14.224449   13.935843   12.680216
#>  [859]   11.865679   13.457649   17.141320   16.540965   11.769932   15.765908
#>  [865]   16.248260   16.636064   14.574708   15.888554   17.160406   13.954159
#>  [871]   13.358337   16.594190   14.464264   16.254945   14.305713   12.244989
#>  [877]   11.875906   17.325393   14.348216   16.185833   15.640195   15.634122
#>  [883]   11.915560   16.923341   15.787302   14.789317   16.484823   12.141801
#>  [889]   14.107653   14.770950   16.974504   17.025532   16.748108   13.919413
#>  [895]   15.869931   17.192139   17.328413   16.564907   15.110308   13.289994
#>  [901]   13.640950   11.894744   12.570383   12.881787   14.510979   16.753175
#>  [907]   16.100521   14.379468   16.180363   15.630111   13.691058   17.085014
#>  [913]   12.476146   13.687195   17.106734   16.649913   13.024962   13.200081
#>  [919]   12.636355   17.087408   16.638481   12.476243   12.177911   14.859381
#>  [925]   15.983883   12.763150   17.055976   11.383724   14.702122   13.173468
#>  [931]   16.582879   15.881117   12.260423   12.958051   16.658632   17.065825
#>  [937]   16.641768   17.271078   13.133473   10.946275   16.963193   16.951665
#>  [943]   15.065543   12.887145   14.560921   10.568852   13.866556   16.482363
#>  [949]    7.254902   12.959414   15.250827   11.827057   16.789396   17.018027
#>  [955]   16.989004   13.408026   16.034216   16.643392   17.282811   17.328149
#>  [961]   16.135322   17.368977   12.634153   10.904038   16.760934   14.571916
#>  [967]   17.085950   12.012618   17.013962   13.580537   16.543195   12.990029
#>  [973]   14.624891   11.049364   11.991077   14.475838   12.936023   16.416938
#>  [979]   12.567368   13.023120   14.985122   14.760887   17.067840   17.144906
#>  [985]   14.028252   12.838215   14.164033   15.798970   16.794868   12.426116
#>  [991]   13.117201   16.655782   16.471549   17.196131   16.540823   15.615520
#>  [997]   16.548512   15.889778   16.222010   17.130135  130.773225   23.026258
#> [1003]   22.572464   23.219468   27.172366   20.460470  283.148973   25.492322
#> [1009]   15.844085   58.563014   13.307410   25.286720   29.653365   52.271003
#> [1015]   58.282609   26.571287   40.311565   17.771131   34.299718   12.407607
#> [1021]   23.016041   58.363826  122.437736   24.985696   54.509682  799.854655
#> [1027]   20.921174   44.613528   41.470033   28.919226   65.275744   14.855115
#> [1033]   20.638212   28.575702   32.566849   40.867738   30.238095   14.944504
#> [1039]   20.082262   36.877179   27.907007   21.701388   35.198340   15.883131
#> [1045]   25.849796   16.484823   30.132223   37.329382   28.325822   21.502392
#> [1051]   19.679472   45.179143   20.131063   71.357818   45.193521   24.944045
#> [1057]   37.832483   55.611332   11.716588   19.685662   17.566947   40.410362
#> [1063]   38.412513   40.564628   26.537921   51.213815   36.566781   21.895167
#> [1069]   21.789082   30.963355   61.173291   37.500393   22.855051   17.987026
#> [1075]   59.829817   58.476649   31.366674   21.126819   28.964492   22.622262
#> [1081]   18.186880   28.530619   86.474881   27.362484   33.776912   48.210073
#> [1087]   18.130365   13.086794   32.301572   33.456151   15.061215   27.133207
#> [1093]   45.629007   26.578739   22.084370   57.877384   42.159573   36.788030
#> [1099]  773.754967   26.799865   20.295727   14.672341   19.782234   23.213424
#> [1105]   36.752655   40.757226   16.300877   33.869815   29.627478   18.028976
#> [1111]   14.306064   63.545944   31.791865   18.420686   27.487953   12.113523
#> [1117]   18.303304   54.998619   43.390526   50.941584   38.544352   38.597736
#> [1123]   23.791014   20.665596   22.127391   76.765101   14.150791   23.981587
#> [1129]   23.201277   51.708421   20.444485   31.430292   19.020747   21.013445
#> [1135]   17.144906   48.264226   20.464977   17.439099   19.259854   15.996836
#> [1141]   62.135850   49.041637   73.515618   44.657457   23.287115   38.834253
#> [1147]   22.560101   18.832072   48.269766   96.842336   26.102624   48.813419
#> [1153]   20.153859   57.517673   43.623372   19.438536   19.710549   23.821340
#> [1159]   20.094795   28.835529   27.232039   30.207733   15.833910   18.481284
#> [1165]   21.222627   20.291659   57.796275   18.300442  277.123033   26.342158
#> [1171]   28.049077   42.531446   18.518450   14.422140   29.992345   24.239482
#> [1177]   26.971368   24.133292   25.394670   19.827351  240.185170  117.286313
#> [1183]   40.406444   40.929668   28.859971   56.441911   44.528961   48.993802
#> [1189]   24.654084   24.621854  137.457248   54.901621   12.998057   36.830120
#> [1195]   63.177814   18.725029   38.514733   21.399751   34.804054   23.420923
#> [1201]   13.166651  274.208668  207.222011   23.676904   25.187301   21.632306
#> [1207]   16.219581   29.465871   45.489532   32.505178   13.609513   36.704834
#> [1213]   25.020624   35.968197   13.715794   16.557493   19.788057   23.227707
#> [1219]   54.961923   15.606316   17.358476   11.378490   50.511990   27.809236
#> [1225]   47.229611   20.371450   17.705525   18.578595   45.526648   18.243357
#> [1231]   30.452709   38.223654   36.931488   17.789530   35.311159   18.695660
#> [1237]   30.273199   15.257498   16.941507   29.426902   22.326598   27.045797
#> [1243]   54.566081   21.305148   53.698449   30.557687   10.570734   52.887508
#> [1249]   30.464560   15.755278   40.198490   50.794935   20.399380   17.997391
#> [1255]   34.268497   16.749688   16.901757   20.015344   40.272936   42.031350
#> [1261]   35.994689   18.714808 1703.874279   13.839600   41.446654   26.796640
#> [1267]   45.179986   44.257704   18.623867   56.375166   65.007020   19.182510
#> [1273]   40.448739   21.858562   21.762969   55.729952   25.301530   54.707903
#> [1279]   14.581040   32.917754   62.658208   30.775669   23.364862   21.218601
#> [1285]   33.293161   13.817897   44.763488   27.193060   35.582202   45.747472
#> [1291]   14.076128   14.612972   24.012403   45.404603   20.630382   23.972281
#> [1297]   31.383094   31.238270   18.195377   13.702181   12.067454   21.563903
#> [1303]   35.259941   17.028259   27.097612  324.902416   11.973104   42.604506
#> [1309]   42.601730   53.219334   19.981856   38.769513   14.286970   25.441025
#> [1315]   48.974786   15.421889   10.575731   24.355682   19.459715   15.841068
#> [1321]   38.834017   52.368287   27.968489   17.793279   22.755586   22.638087
#> [1327]   34.628934   31.406560   14.651815   23.370741   45.001952   40.579967
#> [1333]   27.697175   34.865451   16.286084   57.119658   22.401619   29.909092
#> [1339]   23.765142   19.752690   27.109329   24.117013   45.169734   39.012569
#> [1345]   21.674116   23.712711   29.884327   37.247013  308.238264   16.091911
#> [1351]   10.957588   17.260698   42.266399   24.199599   26.784919   54.470010
#> [1357]   21.446951   40.499077   28.344567   21.476100   24.088721   79.123392
#> [1363]   23.736203   36.026897   66.491230  195.953230   18.349267   36.183700
#> [1369]   15.230496   12.727935   17.163368   14.653257   14.075988   16.328572
#> [1375]   24.938290   18.207747   40.525034   57.904279   16.142675   14.920402
#> [1381]   19.549235   67.642994   41.160680   20.541396   16.180792   15.080599
#> [1387]   27.053093   39.491955  249.080439   47.787861   50.736616   15.497496
#> [1393]   26.555819   21.747596   16.673937   21.145785   68.629986   40.899390
#> [1399]   13.752505   24.842969   43.455548   38.064746   52.022451  634.639976
#> [1405]   29.758722   29.114687   19.462434   53.969363   46.499995   44.122996
#> [1411]   46.442612   20.573222   54.576806   17.933390   13.640950   25.245039
#> [1417]   44.973426   18.048183   19.061858   12.479169   26.559577   18.676677
#> [1423]   18.494245   22.337346   26.060711   45.711215   12.384606   39.483191
#> [1429]   39.214219   14.985122   13.530286   59.822127   29.270672   60.696587
#> [1435]   17.873633   24.439408   43.401628   27.256983   17.733810   21.834093
#> [1441]   22.861844   36.163742   28.168580   28.999739   46.165548   16.700328
#> [1447]   18.033277   27.264321   28.471868   20.852569   21.731533   25.541403
#> [1453]   34.965903   42.259171   24.632911   26.910043   34.157840   13.777360
#> [1459]   27.916243   13.598537   34.981977   22.129830   18.130833   14.507506
#> [1465]   40.400493   22.711987   41.763965   19.127325   31.497081   13.960203
#> [1471]   20.543129   12.768662   23.205481 1123.788164   20.094317   26.183340
#> [1477]   32.704051   26.382949   48.151032   38.372986   30.680708   42.878495
#> [1483]   33.237893   18.194070   46.147022   10.696309   58.114951   15.461467
#> [1489]   22.195249   59.755399   13.654239   13.655725   42.260679   68.159228
#> [1495]   15.888552   19.863588   23.122169   18.978064   19.326554   30.236075
#> [1501]   29.208765   53.671537   72.041909   16.520802   15.890825   34.088152
#> [1507]   55.710484   34.151001   39.791001   33.542550   12.424456   28.613153
#> [1513]   35.628642   25.407532   20.026231   20.492405   34.044493   19.850730
#> [1519]   21.887294   42.270349   20.057446   21.173006   93.728684   50.375863
#> [1525]   29.629268   23.042936   27.540867   24.525850   21.929180   20.546081
#> [1531]   55.452224   27.483339   51.331254   29.866362   32.275359   24.882029
#> [1537]   43.409795   34.704551   15.336005   17.389455   12.611562   41.164443
#> [1543]   17.975210   27.655653   47.784089   24.059236   33.411425   72.703438
#> [1549]   28.948863   21.350492   34.392755   25.662696   17.165878   16.847252
#> [1555]   55.171339   50.593714   15.046025   49.980016   44.405306  582.615073
#> [1561]   20.393996   19.471928   20.595717   10.003190   16.224718   56.530657
#> [1567]   41.560041   29.518041   38.274692   27.861566   45.613134   14.566770
#> [1573]   20.324419   63.181043   18.114791   22.653713   12.936023   26.207105
#> [1579]   26.507350   12.623086   27.528525   49.187193   19.495050   45.649954
#> [1585]   29.986491   41.909861   24.275498   57.655474   54.374986   19.608471
#> [1591]   35.409148   17.231882   16.128101   95.222236   37.339863   29.878574
#> [1597]   22.235358   26.657028   21.174532   36.688038   29.650868   16.689944
#> [1603]   17.471494   25.031802   22.949773  160.439337   22.393292   10.568852
#> [1609]   32.115330   10.723499   51.276413   15.903981   19.986044   16.055788
#> [1615]   21.719764   49.359082   22.448363   35.141238   27.471314   19.392380
#> [1621]   21.051269   14.883068   28.318769   15.771837  224.589966   21.309370
#> [1627]   18.328840   38.321143   27.452416   15.183044   16.644758   30.682340
#> [1633]   40.946686   19.588947   18.994665   17.860170   17.887232   16.933400
#> [1639]   19.388153   51.158559   15.032992   20.568417   20.208904   12.788164
#> [1645]   17.440530   47.245426   30.051539   16.262581   37.323506   42.166024
#> [1651]   28.450956   31.694298   42.072211   17.942198   13.775429   29.289657
#> [1657]   24.168448   21.239452   16.397520   15.718205   19.831474   25.954654
#> [1663]   20.637245   21.698901   17.410348   41.575101   18.514720   26.342532
#> [1669]   26.666621   14.165024   16.546013   30.484583   17.665470   36.211792
#> [1675]  181.173954   35.618206   26.080338   17.477554  388.846253   24.649500
#> [1681]   32.099081   55.067202   36.561741   26.176128   35.066080   25.181493
#> [1687]  571.109703   51.165378   25.856113   15.815677   14.418261   50.450245
#> [1693]   16.334002   28.846182   21.255140   21.278368   65.175524   20.008403
#> [1699]   19.702259   33.102467   18.149483   21.926406   40.172012   12.457246
#> [1705]   20.109581   40.452812   12.838215   35.820362   21.171141   42.656180
#> [1711]  113.502567   13.924657   19.914275   16.687774   21.241118   29.934532
#> [1717]   29.325576   34.394410   41.671014   22.362791  241.828638   24.359829
#> [1723]   17.215016   44.107981   24.407078   27.169102   39.050123   18.891613
#> [1729]   24.176613   18.269880   22.338860   26.341217   50.688338  392.404297
#> [1735]   27.407282   23.048269   34.518468   26.124329   15.817078   24.390033
#> [1741]   12.501835   34.653892  610.129584   42.971134   25.024345   34.681646
#> [1747]   44.685687   36.563414   12.052644   13.058699   36.650341   15.065388
#> [1753]   32.253745   21.157596   24.886527   19.129110   16.504925   19.572174
#> [1759]   29.462565   25.756343   25.151111   43.670847   59.298541   23.566054
#> [1765]   30.911250   48.897875   29.672442   18.315660   17.434028   27.256940
#> [1771]   49.036971   45.555449   12.763150   63.029778   31.336875   46.907507
#> [1777]   22.189718  360.422301   39.575329   24.555672   15.307088   23.402203
#> [1783]   12.692997   19.526947   29.000534   64.223715  629.484846   27.641833
#> [1789]   60.191119   12.553355   41.303662   27.635110   37.306598   28.755265
#> [1795]   20.382917   20.315644   14.905035   68.374453   39.337244   28.912131
#> [1801]   91.249954   20.975506   46.674186   13.830143   26.563378   15.833526
#> [1807]   19.620317   18.398684   12.390612   53.238678   32.298563   44.405148
#> [1813]  141.001997   42.338535   23.672778   16.454342   17.317236   23.452802
#> [1819]   40.641884   51.661313   16.306120   22.655073   40.111795   17.650548
#> [1825]   21.551384   32.164458   36.101689   27.552572   54.054065   51.585943
#> [1831]   31.968951   38.742837   26.326390   46.975668   33.441781   25.518472
#> [1837]   13.597145   44.153706   19.113122   29.454299   20.463728   27.595930
#> [1843]   37.834033   22.355775   17.750134   19.571854   60.929024   13.213845
#> [1849]   18.163496   41.791428   14.916574   55.330184   29.300358   24.270834
#> [1855]   13.436431   38.145838   46.096178   26.142182   27.659140   11.273354
#> [1861]   21.968514   16.213107   25.979802   17.083093   26.268958   42.868691
#> [1867]   19.378582   24.198304   18.482662   51.086388   18.916058   92.389747
#> [1873]   16.646448   20.207058   53.785728   37.274327   21.996297  213.225993
#> [1879]   15.081716   22.704373   44.194044   13.456953   17.847863   94.630726
#> [1885]  187.662173   21.542231   30.667803   33.523851   31.100062 1391.225763
#> [1891]   40.670285   27.123212   15.017940   36.846253   17.786141   84.311469
#> [1897]   20.870218   17.485338   15.110308   21.581876   29.619270   46.759872
#> [1903]   56.188513   31.365107   28.184812   13.960745   35.503690   54.047793
#> [1909]   43.187814   13.788915   12.698473 1951.368968   21.313444   66.236713
#> [1915]   13.387085   44.083321   43.276610   39.320420   41.081116   91.721243
#> [1921]   17.477809   22.260098   11.827606   25.211763   20.208744   14.268100
#> [1927]   23.877979   33.752868   50.573232   18.230693   17.149547   39.767031
#> [1933]   21.804286   39.827044   28.537966   46.344788   13.447009   34.456436
#> [1939]   49.609262   12.504358   46.280491   18.775832   12.701609   17.824774
#> [1945]   24.368959   51.798234   13.826561   20.381482   49.940858   25.786892
#> [1951]   21.473197   36.335343   91.673162   23.824948   30.062635   22.535872
#> [1957]   35.116634   25.650960   48.251100   44.371265   16.558264   19.250396
#> [1963]   23.146223   21.996351   88.180360   45.227528   15.775518   23.455499
#> [1969]   19.173187   56.217312   22.003501   36.703551   15.428994   16.970441
#> [1975]   40.335842   20.301629   36.304582   30.942527   17.796548   18.149582
#> [1981]   16.181928   12.994569   17.553082   39.987802   42.113296   14.889066
#> [1987]   35.098675   29.320344   24.447373   16.014765   20.559788   27.448910
#> [1993]   30.376041   52.295043   32.650436   35.844525   16.611294   22.418915
#> [1999]   21.089923   34.629608