Long-term stochastic growth rate worker function for the xsdm model

Computes the negative of the long-term stochastic growth rate, plus log(lambda_max), for the xsdm model, for each location. This is the fast version of a worker function that relies on C++ code, optimized using RcppParallel. The legacy pure-R reference is preserved as xsdm:::like_neg_ltsgr_r for testing and comparison.

Usage

like_neg_ltsgr(
  env_dat,
  mu,
  sigltil,
  sigrtil,
  o_mat,
  num_threads = RcppParallel::defaultNumThreads()
)

Arguments

env_dat

The environmental data array, dimensions (number of locations) x (time series length) x (number of environmental variables). Must not contain missing values.

mu

Vector of optimal environmental values. Length p=dim(env_dat)[3]. Unconstrained real numbers.

sigltil

Vector specifying width of the growth-environment function. Length p=dim(env_dat)[3]. Positive real numbers, Inf entries also allowed.

sigrtil

Vector specifying width of the growth-environment function. Length p=dim(env_dat)[3]. Positive real numbers, Inf entries also allowed.

o_mat

An orthogonal matrix, dimensions p by p.

num_threads

Number of threads for parallel computation. Defaults to RcppParallel::defaultNumThreads().

Value

A vector of length equal to the number of locations, as described above.

Details

Internally, this function:

Reshapes the environmental data into a matrix.

Computes inverse (diagonal) matrices for asymmetry adjustments.

Calls the C++ function like_ltsg() for efficient likelihood computation.

Note

Ensure that env_dat has no missing values. The parameter vectors mu, sigltil, and sigrtil must have length equal to the number of environmental variables (p). The argument o_mat must be a p x p orthogonal matrix (i.e., o_mat %*% t(o_mat) is the identity).

Examples

# Example usage:
like_neg_ltsgr(env_dat = example_1$env_array,
               mu      = example_1$par_list$mu,
               sigltil = example_1$par_list$sigltil,
               sigrtil = example_1$par_list$sigrtil,
               o_mat   = example_1$par_list$o_mat)
#>    [1]   2.738171   8.535266   8.374321   7.154807   8.444485   3.477033
#>    [7]   3.272479   8.034020   5.494181   3.406080   3.894184   8.221686
#>   [13]   7.075254   5.269100   4.288629   3.740807   6.999191   4.930312
#>   [19]   6.834511   2.327886   2.074068   4.776596   2.464446   6.709562
#>   [25]   8.304522   3.820160   8.435004   5.298979   6.876833   4.155158
#>   [31]   6.926461   6.161269   4.097653   5.102134   4.291076   6.345642
#>   [37]   4.245375   3.341317   8.660052   6.664898   2.810356   5.362398
#>   [43]   6.999660   7.627527   4.651365   5.483883   4.516378   5.404858
#>   [49]   3.790042   6.025831   8.251108   6.783180   7.984423   4.287914
#>   [55]   3.056667   6.856656   8.214739   7.498210   5.261167   5.338618
#>   [61]   8.011760   5.193644   4.798711   8.416415   7.293172   3.600957
#>   [67]   4.264491   8.319323   6.217450   4.367401   5.124876   1.979788
#>   [73]   7.917596   6.576324   3.495894   5.479428   7.474645   6.975485
#>   [79]   3.053595   3.440059   4.145340   4.927784   3.322202   4.472423
#>   [85]   6.570589   7.014075   5.538776   2.680407   4.693739   6.103064
#>   [91]   6.448714   2.783318   5.408007   8.305771   3.643784   6.221024
#>   [97]   6.240697   7.991659   5.309896   3.645075   4.120219   8.335431
#>  [103]   5.147934   3.522013   8.410355   4.912260   4.622131   5.170946
#>  [109]   7.813666   7.410260   7.780804   7.690245   4.278769   4.433976
#>  [115]   6.439727   4.491546   5.024642   2.591280   7.766783   5.951452
#>  [121]   6.646883   7.701346   3.647620   7.711426   4.535503   6.205746
#>  [127]   4.217129   6.303538   4.647464   8.543404   7.927106   7.634204
#>  [133]   6.767605   5.450163   5.276290   8.677979   4.104554   6.750432
#>  [139]   5.101582   3.535671   4.327127   3.899442   2.513490   6.251582
#>  [145]   4.136154   8.488412   7.446858   2.584259   6.671945   3.851036
#>  [151]   7.779567   8.397149   3.530302   5.730067   7.852212   4.845694
#>  [157]   4.944500   5.696837   8.223739   4.374859   5.328191   7.272478
#>  [163]   3.679817   6.775375   8.134278   5.630475   7.462363   8.346814
#>  [169]   5.211055   5.127361   4.745629   7.709655   6.691052   2.921612
#>  [175]   6.464432   8.342831   3.845642   8.447241   6.873000   5.316705
#>  [181]   6.203020   2.769259   7.306254   2.951352   8.288232   3.652701
#>  [187]   6.782896   7.864908   6.941641   7.954834   7.279754   7.033723
#>  [193]   3.971800   3.557843   3.819581   5.665545   8.269367   8.213562
#>  [199]   4.699454   5.075297   7.673952   2.967565   5.479933   3.740539
#>  [205]   7.866173   5.794756   3.110880   6.288045   4.136789   5.034436
#>  [211]   3.201250   3.983400   7.933329   6.077172   7.230290   6.730007
#>  [217]   5.377706   5.802934   3.670574   2.933928   5.957608   7.205059
#>  [223]   3.711949   6.983483   3.239194   6.335887   3.289620   6.014989
#>  [229]   4.945292   3.322412   3.503462   7.155621   5.889508   3.292937
#>  [235]   8.294997   2.559470   3.128033   3.970428   5.658968   5.905711
#>  [241]   2.615392   2.456258   3.667483   5.087955   7.433216   7.586851
#>  [247]   7.706441   5.496604   6.784210   4.197012   3.969127   7.780620
#>  [253]   4.569003   3.739359   7.585584   6.283181   2.361769   7.290611
#>  [259]   6.927343   7.520906   8.527811   6.068080   7.204822   4.434878
#>  [265]   2.492797   8.022582   8.027586   4.773539   6.594982   2.860595
#>  [271]   7.013509   8.685222   6.542163   2.561165   2.586524   6.109590
#>  [277]   2.760930   8.026780   3.050958   1.844593   8.163568   3.374663
#>  [283]   6.314393   7.206267   5.450465   6.726003   7.282657   7.573248
#>  [289]   4.536227   7.985194   4.070292   4.351968   5.404378   5.499150
#>  [295]   5.836856   6.364855   3.162164   5.752320   6.161986   7.273024
#>  [301]   5.235339   8.622596   2.973380   3.124899   2.943223   6.379983
#>  [307]   6.403187   7.425073   8.422545   5.931950   6.899647   6.816617
#>  [313]   6.971037   3.377473   4.321292   3.693707   4.793186   3.900874
#>  [319]   2.681686   8.605958   4.206489   7.616487   7.673731   7.714181
#>  [325]   6.153605   4.510407   2.689889   6.279882   4.343616   6.426235
#>  [331]   6.087054   3.416215   7.227908   4.843000   2.985486   2.792633
#>  [337]   3.307358   7.258635   7.551705   3.969658   4.025248   5.751400
#>  [343]   2.035704   6.497565   7.933129   5.234216   5.687010   2.539760
#>  [349]   5.279772   5.829322   6.294680   3.272314   7.152536   6.923463
#>  [355]   6.968193   6.619868   3.948832   8.679205   6.135075   5.950581
#>  [361]   3.456851   4.783452   3.620395   4.837772   6.228984   3.439286
#>  [367]   6.268120   3.927663   2.447301   7.229603   7.322477   3.821366
#>  [373]   7.705315   8.036480   5.540306   7.080314   5.447749   5.095901
#>  [379]   2.527407   6.584077   3.508366   4.682486   4.101918   3.799091
#>  [385]   4.973119   3.985040   3.729906   5.554829   7.094541   8.019271
#>  [391]   2.329663   6.167402   5.599354   3.247175   5.518319   8.286294
#>  [397]   8.671559   8.665080   3.020188   4.232752   3.451821   3.451614
#>  [403]   8.401605   2.717455   7.915186   6.504449   2.680351   7.584430
#>  [409]   8.101095   7.715503   7.358174   3.877302   3.030461   2.558614
#>  [415]   6.942178   4.847294   4.189646   3.237823   5.282576   3.985766
#>  [421]   2.112286   7.474094   3.878193   5.340085   6.494366   4.547283
#>  [427]   5.160111   6.072133   2.160651   3.737604   4.948909   1.936117
#>  [433]   7.845225   2.615043   5.006440   6.472571   8.186124   4.570254
#>  [439]   8.066034   3.448542   7.934359   7.895704   6.997027   7.890756
#>  [445]   6.198348   3.660974   4.823904   7.650551   7.833773   4.227792
#>  [451]   6.156741   4.733819   8.015880   4.603286   3.545889   6.032412
#>  [457]   4.575272   6.251188   7.822554   4.084764   5.621771   3.505276
#>  [463]   4.075021   5.244029   7.060772   5.471940   6.861719   4.819562
#>  [469]   2.956816   7.756185   8.016243   2.872487   4.123046   2.986877
#>  [475]   6.151864   5.009757   7.639867   3.309031   3.169437   6.796955
#>  [481]   5.730892   3.721554   8.480516   6.774152   4.392391   3.517795
#>  [487]   6.480300   6.207808   4.784948   2.357236   4.293837   5.003840
#>  [493]   8.260322   6.167962   6.685890   5.986726   4.439560   5.268185
#>  [499]   3.692445   4.995306   8.440351   6.064291   7.957908   5.430812
#>  [505]   7.770940   8.281199   3.646238   6.410633   7.655068   5.583325
#>  [511]   7.590048   5.729619   3.557723   5.335095   3.581894   3.319113
#>  [517]   6.545221   7.913582   8.127177   5.752468   7.685988   5.670282
#>  [523]   4.941975   5.848321   3.864835   3.503714   3.982686   7.273977
#>  [529]   3.939822   4.406699   5.671054   7.422249   3.043856   5.786128
#>  [535]   2.259710   5.274936   5.710916   5.117274   5.661057   4.056405
#>  [541]   4.590258   7.912187   4.334486   7.912549   2.525762   8.063547
#>  [547]   4.318433   4.853824   4.865807   7.624448   5.712030   5.489711
#>  [553]   3.967668   4.848766   6.506274   5.443358   8.577216   7.635640
#>  [559]   5.873814   6.744487   4.973482   2.775246   8.286285   7.922041
#>  [565]   8.471732   4.307777   8.196540   6.993114   4.157498   7.781853
#>  [571]   7.558540   3.075448   4.393681   7.980965   4.469846   4.658037
#>  [577]   5.480192   3.430931   4.090203   3.912564   2.638866   3.212021
#>  [583]   5.162167   5.341303   5.499401   5.847025   4.660721   3.823180
#>  [589]   5.767246   7.377622   2.930334   3.374140   5.982484   2.714523
#>  [595]   7.547959   7.244319   5.609992   4.784291   6.344278   7.084156
#>  [601]   4.802776   4.274909   6.449187   3.427608   3.693117   4.069509
#>  [607]   5.456143   3.669372   5.619555   5.913603   8.476933   3.498282
#>  [613]   5.432434   8.408024   8.596682   3.182762   4.300191   6.159364
#>  [619]   4.843586   4.841264   5.370967   3.821050   4.362531   7.689048
#>  [625]   6.665243   7.154496   4.033387   3.555588   4.593265   3.377781
#>  [631]   6.919569   7.298779   5.422203   4.555155   6.737008   4.172477
#>  [637]   2.553697   6.720496   4.164538   2.596982   8.347730   2.785452
#>  [643]   8.259568   4.302055   2.352061   4.563356   4.646854   6.553449
#>  [649]   7.443764   7.206827   6.304581   4.618992   3.244508   8.018727
#>  [655]   4.766374   4.769661   5.961189   4.817558   7.126731   4.181735
#>  [661]   6.490961   5.532172   7.067334   3.295478   7.676368   2.496864
#>  [667]   5.300190   4.988234   7.448746   7.893728   6.219486   6.874343
#>  [673]   5.505493   7.168542   8.511330   3.726587   3.764189   5.913508
#>  [679]   4.939287   3.778838   4.851821   8.122428   2.916222   6.696599
#>  [685]   6.428615   4.561121   5.282771   6.382401   4.725508   6.023216
#>  [691]   4.346883   3.913192   5.304983   6.179082   6.734587   4.982478
#>  [697]   5.648537   5.677169   3.188524   8.595754   5.435996   7.867928
#>  [703]   6.716072   8.012291   6.670052   4.191274   5.503510   4.450612
#>  [709]   6.247518   3.287386   3.797807   2.658075   3.285803   6.342421
#>  [715]   5.439417   6.993322   6.756449   4.612762   3.195821   5.302823
#>  [721]   5.575945   5.568095   7.792703   6.585643   7.483930   4.074094
#>  [727]   5.666728   7.088191   3.122774   6.049405   2.409910   4.090649
#>  [733]   4.903148   7.570804   8.474624   4.770945   4.154927   5.334148
#>  [739]   6.734582   4.504836   4.523618   3.278765   5.442380   3.066054
#>  [745]   5.866659   6.170346   3.808898   6.223310   7.027555   8.021163
#>  [751]   7.068678   4.851655   7.773158   7.231622   8.090406   3.312508
#>  [757]   8.166651   8.480308   3.025115   5.567738   7.806375   8.507377
#>  [763]   4.539791   3.231424   2.704578   7.825126   6.217286   6.713378
#>  [769]   5.634191   6.494820   3.397384   4.868878   3.726779   5.389809
#>  [775]   4.076884   4.277643   3.672147   3.998512   8.021931   7.838774
#>  [781]   3.904019   5.091911   6.722692   6.083745   5.248950   8.037602
#>  [787]   3.238276   5.888718   4.236371   5.590665   8.668876   3.990844
#>  [793]   4.194881   8.379268   3.494161   3.836587   7.120971   7.651745
#>  [799]   5.233000   7.799905   4.267790   4.726320   7.565624   2.667601
#>  [805]   4.301659   5.626010   3.534682   7.736772   2.358384   6.294209
#>  [811]   2.887388   4.085603   2.767285   3.562569   2.754517   4.961691
#>  [817]   5.095786   6.523526   8.224669   6.337783   8.053719   4.533643
#>  [823]   3.175342   2.494171   5.932482   8.085847   4.204271   5.116782
#>  [829]   2.709875   5.831805   2.735037   7.964228   6.569012   8.115789
#>  [835]   3.135490   8.110443   7.832184   3.689198   7.191854   3.034071
#>  [841]   2.842213   7.979132   3.830341   5.214407   3.381384   8.616595
#>  [847]   5.219445   8.271040   3.862106   7.888417   6.245119   4.562539
#>  [853]   3.699590   2.962022   7.240975   7.161624   5.060736   7.313176
#>  [859]   3.288401   5.885004   4.564450   7.303013   6.814671   8.475478
#>  [865]   7.860754   8.451341   4.688513   4.678283   5.727080   5.170603
#>  [871]   4.291227   6.475604   2.603295   2.760989   7.459558   8.102027
#>  [877]   7.461878   2.410126   7.807386   7.194237   6.086081   3.552605
#>  [883]   7.114467   7.744252   8.519727   8.272525   5.014570   4.142912
#>  [889]   4.224112   6.343662   2.820095   2.939107   5.143522   8.191626
#>  [895]   7.272415   4.223209   7.604237   7.504105   4.957457   2.690517
#>  [901]   4.707225   5.840925   6.673861   7.387350   3.126374   6.841601
#>  [907]   8.469149   5.023942   6.834962   8.106575   6.921475   3.870649
#>  [913]   5.548968   7.394421   7.157734   3.242620   4.951736   4.804114
#>  [919]   6.130948   6.544578   4.533288   5.959637   8.362821   4.600947
#>  [925]   6.755693   2.414515   4.478275   3.900118   6.463981   6.301487
#>  [931]   5.571437   8.191842   8.676462   8.067420   2.158200   5.778746
#>  [937]   5.974072   6.132136   8.461567   3.152480   7.282938   6.736813
#>  [943]   6.954803   5.222599   8.590609   8.580312   6.749103   4.947143
#>  [949]   5.073389   4.508214   8.585070   3.383786   7.074032   4.921445
#>  [955]   4.155504   5.249568   6.704727   4.163071   4.054647   5.207599
#>  [961]   5.223112   6.998289   3.515707   6.064597   7.702133   8.125316
#>  [967]   8.536342   6.053649   3.477018   6.206682   6.722753   2.224025
#>  [973]   8.178045   4.694019   8.346858   6.135494   3.704659   5.976697
#>  [979]   5.085363   3.403459   8.200203   5.356457   6.778927   4.348040
#>  [985]   6.057429   4.620366   7.120062   3.405037   2.839946   2.746828
#>  [991]   2.727760   3.774420   3.470485   3.305958   4.103848   5.966980
#>  [997]   7.354790   6.200861   4.372938   7.222016   7.995522   2.706947
#> [1003]   5.657081   7.401345   5.180398   6.017615   3.397586   3.838543
#> [1009]   3.307787   3.724071   4.864723   8.434487   3.470315   2.221633
#> [1015]   5.291677   4.677225   6.028211   8.593110   3.175965   8.499105
#> [1021]   8.478475   7.085858   4.073289   8.657515   4.282575   4.028622
#> [1027]   2.480990   7.586578   4.139562   3.200358   5.155334   8.022828
#> [1033]   3.960390   3.780269   8.667614   8.385096   8.607728   8.591794
#> [1039]   5.421297   5.397384   7.761876   5.952748   2.583469   3.831588
#> [1045]   8.374650   8.218285   3.749434   2.872873   7.308798   8.289128
#> [1051]   8.216007   5.660529   8.062701   7.871427   7.349598   7.606838
#> [1057]   8.378988   6.421674   6.442160   6.725342   6.499535   5.917280
#> [1063]   7.511881   6.785173   5.170750   5.050956   4.664727   7.380811
#> [1069]   6.599963   4.220824   6.574463   8.131187   6.929954   3.010809
#> [1075]   5.410238   6.390481   3.090726   3.521142   5.864787   8.399053
#> [1081]   5.146584   3.115164   7.786884   7.777974   7.078851   6.944879
#> [1087]   5.328855   5.598021   3.689548   5.751970   7.624694   7.375325
#> [1093]   8.318286   7.113092   3.795615   5.720884   3.330763   5.135225
#> [1099]   7.433924   6.436189   3.499733   6.407128   3.527132   4.261679
#> [1105]   8.200243   7.436496   5.884981   5.818474   3.214951   2.911497
#> [1111]   7.444661   3.338303   8.544388   3.384265   6.328935   7.463535
#> [1117]   2.707121   8.138389   2.718702   5.612178   4.285213   3.885040
#> [1123]   3.217582   4.595973   3.357771   4.668416   4.714070   3.329627
#> [1129]   6.255171   3.250422   6.792938   4.970029   6.992931   8.052975
#> [1135]   8.434174   5.006356   4.567673   2.656590   2.052505   5.623835
#> [1141]   4.736700   5.225845   3.326151   5.385088   6.092414   4.702879
#> [1147]   4.383010   2.966236   7.639685   5.980486   7.461214   3.645708
#> [1153]   4.737205   2.019887   8.580695   6.665618   6.299538   3.101122
#> [1159]   3.653575   2.747576   7.087695   6.686088   7.159726   6.589873
#> [1165]   4.615975   7.699686   7.103967   8.582476   2.993424   3.762590
#> [1171]   5.092544   2.747470   4.703475   8.600092   4.408565   5.324506
#> [1177]   8.324865   4.010067   7.045706   2.495274   4.268073   6.553449
#> [1183]   4.699161   6.141461   3.733548   6.921081   7.167500   5.260439
#> [1189]   3.020323   7.027039   8.582762   4.055621   7.205496   6.228757
#> [1195]   8.498027   3.515893   5.464716   4.983289   7.179643   3.493865
#> [1201]   5.427163   8.055494   7.842102   5.207980   2.739558   4.182599
#> [1207]   3.897595   6.621798   5.774103   8.108964   6.626362   3.184283
#> [1213]   6.190446   4.582555   7.796147   5.600708   5.053921   5.048948
#> [1219]   7.824117   3.027946   4.861814   5.391532   7.955225   5.204822
#> [1225]   8.645558   3.972073   7.733587   5.407027   3.629083   4.631664
#> [1231]   4.677387   3.039714   6.103555   6.411985   3.117300   6.210034
#> [1237]   7.593603   7.385015   5.499364   7.157728   7.929159   4.298047
#> [1243]   6.374831   3.492197   2.707568   5.192167   3.450740   5.069526
#> [1249]   2.994888   5.014392   5.580315   8.443390   6.388729   4.689204
#> [1255]   7.584902   4.066775   7.449966   2.520724   7.062368   7.199670
#> [1261]   7.263620   3.116505   4.997896   1.997175   5.051030   4.279090
#> [1267]   5.356132   3.791777   6.143416   8.115884   7.929082   7.003615
#> [1273]   6.733263   4.934277   6.385127   8.024584   6.579481   6.182975
#> [1279]   5.077234   3.600283   7.880312   3.515523   2.784094   8.649873
#> [1285]   6.532410   8.268666   6.753092   3.134213   5.103976   4.831819
#> [1291]   5.001506   8.099398   3.799173   7.028379   3.616465   6.303974
#> [1297]   5.892544   5.771123   5.484970   3.336990   2.442142   6.380442
#> [1303]   5.486077   6.440498   7.830177   3.597067   5.110941   3.150174
#> [1309]   3.460901   5.782609   5.905815   2.884812   8.539653   6.069464
#> [1315]   5.574730   4.216050   6.649858   5.537416   4.908490   7.565509
#> [1321]   4.001298   3.523788   7.709303   6.167361   7.757152   3.364267
#> [1327]   7.989933   8.103349   8.393792   3.880307   4.259917   5.610198
#> [1333]   5.939607   2.874106   5.161961   3.084476   3.541698   5.983427
#> [1339]   3.305192   7.062382   8.383849   6.312975   5.185237   3.705460
#> [1345]   5.359076   7.074372   3.712219   5.774309   4.015924   6.900965
#> [1351]   5.048162   7.507350   8.407535   3.404963   4.701133   4.639239
#> [1357]   7.855582   5.946391   8.311488   5.294310   7.969380   5.489818
#> [1363]   4.666057   5.110654   5.549308   8.453464   4.073787   6.308936
#> [1369]   3.375351   4.334426   7.417192   5.328752   8.203106   3.126129
#> [1375]   4.714439   6.047819   3.271284   4.841475   6.185101   3.094741
#> [1381]   5.273980   6.786864   5.688143   6.142299   7.099397   5.447591
#> [1387]   2.708801   5.370271   3.483947   6.596456   3.183036   4.200981
#> [1393]   8.082005   8.475528   5.202435   7.457621   7.169441   7.757026
#> [1399]   7.804030   6.875893   5.683311   7.622484   4.181562   8.404038
#> [1405]   6.554315   8.056132   5.472600   3.588997   5.157280   5.071423
#> [1411]   6.242202   6.267557   5.027541   6.790209   5.809742   5.195028
#> [1417]   8.653050   4.947723   5.312708   3.078897   6.561312   5.622764
#> [1423]   8.444220   8.422796   3.345116   6.922942   4.552223   2.995395
#> [1429]   5.779763   4.580608   2.372507   2.777688   8.167705   7.462094
#> [1435]   6.528297   7.427563   3.453970   4.724074   7.525885   8.596874
#> [1441]   3.275816   5.542566   6.091589   5.212785   2.422029   3.852957
#> [1447]   4.308306   5.607879   2.317805   7.958173   4.548817   6.618372
#> [1453]   3.493687   3.960206   6.142874   4.386309   4.892176   2.872400
#> [1459]   3.324640   7.394688   7.988043   2.965375   1.926234   4.713117
#> [1465]   2.418845   6.723739   7.807502   5.084494   8.132952   7.770551
#> [1471]   6.947474   6.491695   3.442680   7.789901   7.291153   5.855414
#> [1477]   5.439613   5.832337   8.098033   2.651232   2.461777   7.084812
#> [1483]   5.372123   6.559981   5.738617   3.102216   6.495852   6.148476
#> [1489]   2.925299   5.685233   5.968438   5.744250   4.927237   6.954032
#> [1495]   4.851439   3.496134   6.134233   7.180180   7.486868   7.123895
#> [1501]   3.600819   7.463669   7.611982   6.913005   6.020763   8.406499
#> [1507]   8.366950   4.005938   3.029608   8.075166   7.735008   3.220802
#> [1513]   7.438807   3.918075   4.882952   6.041438   6.447468   7.606291
#> [1519]   6.631856   2.885369   6.207476   4.605052   5.161524   8.369215
#> [1525]   8.462632   5.681607   7.511520   7.066230   4.506690   8.465844
#> [1531]   7.379651   7.035396   4.736381   7.436013   6.141569   8.672552
#> [1537]   8.401464   6.157741   2.404245   5.997320   5.296319   5.048998
#> [1543]   4.682357   4.013185   5.680281   3.818874   7.074763   6.910908
#> [1549]   3.434490   4.108423   3.674738   5.819609   5.462443   7.605317
#> [1555]   3.463841   6.284573   5.774066   5.945263   3.417574   5.244412
#> [1561]   7.031786   4.054288   4.983005   8.383029   3.804539   4.919287
#> [1567]   3.345260   5.153427   7.781386   6.929798   7.920586   2.413942
#> [1573]   7.772086   5.774754   6.083392   3.684691   6.302229   3.985111
#> [1579]   2.804628   8.131075   5.646549   7.597837   8.072224   6.831816
#> [1585]   8.280306   2.745323   8.108678   3.615241   3.546137   3.735960
#> [1591]   3.278781   5.658764   3.053661   6.735989   5.012276   7.585056
#> [1597]   6.407779   7.693560   5.667522   8.390397   6.600109   4.323266
#> [1603]   3.587504   3.457487   8.579661   5.571464   6.841271   8.257442
#> [1609]   6.555964   5.500250   5.159479   6.642817   8.269812   7.529613
#> [1615]   5.806236   2.146022   7.197179   8.249122   5.188054   5.027513
#> [1621]   4.043686   4.989440   8.176165   3.611606   7.287544   4.838891
#> [1627]   7.957141   8.046192   3.923782   4.746346   5.759484   2.699770
#> [1633]   6.894886   5.776611   3.758877   8.350999   6.214009   4.982026
#> [1639]   5.777626   6.998908   6.297229   6.078838   8.446107   3.848147
#> [1645]   6.909680   6.816737   8.599369   4.620082   6.823321   5.688064
#> [1651]   8.209079   3.835937   7.785986   7.495871   3.338369   3.007907
#> [1657]   3.481462   3.274610   7.740601   7.607018   8.350273   4.018617
#> [1663]   4.686334   3.705337   6.328251   2.983106   7.573269   4.308838
#> [1669]   8.595109   5.394256   8.507963   6.054456   6.927444   6.850989
#> [1675]   2.751096   5.960060   3.434915   5.359558   7.508456   3.520063
#> [1681]   6.720215   6.927523   4.328956   5.198368   7.542512   2.495729
#> [1687]   7.330701   3.817522   7.681492   5.878555   5.657894   5.227938
#> [1693]   5.212661   4.931369   8.191861   4.986764   3.202317   7.112227
#> [1699]   6.368802   6.466335   4.524100   8.544431   6.349496   5.726599
#> [1705]   6.003285   6.859443   7.163393   8.432526   6.694062   6.074015
#> [1711]   3.545512   7.964937   6.713246   3.472913   5.854436   3.783799
#> [1717]   5.111266   8.043918   6.902310   5.249138   3.083687   6.361188
#> [1723]   8.491980   5.978200   4.787552   4.177960   5.417662   6.779988
#> [1729]   7.155442   2.743534   5.192749   6.640334   7.131507   3.413446
#> [1735]   5.620328   3.174648   3.783612   1.968028   6.277021   3.470196
#> [1741]   4.503302   6.873081   3.547828   7.825469   5.278206   5.601873
#> [1747]   6.716790   7.780396   4.928149   2.817033   7.759968   5.619550
#> [1753]   6.504367   6.971950   6.086410   3.685212   5.721177   4.423524
#> [1759]   6.355705   8.592941   2.908463   4.236645   7.722776   8.384522
#> [1765]   6.457321   3.994591   5.161659   4.732223   7.011291   6.376558
#> [1771]   8.372960   7.116385   5.385134   6.131482   6.202068   6.229191
#> [1777]   2.820421   6.115697   3.178949   5.976989   7.834478   7.426534
#> [1783]   6.236316   8.611529   8.081423   6.412783   5.509415   8.196256
#> [1789]   3.517921   2.453921   8.176636   4.426919   8.450843   5.527404
#> [1795]   5.201377   5.977052   4.878783   4.347502   6.824898   3.798096
#> [1801]   3.186223   5.929647   5.124010   3.241811   5.205021   7.554283
#> [1807]   5.723836   8.092929   6.236936   3.670086   5.343188   3.931252
#> [1813]   2.183416   6.253494   6.001313   5.204241   5.220890   8.005666
#> [1819]   4.044443   5.593127   7.865249   3.407620   1.940691   8.489842
#> [1825]   4.780678   6.727413   3.068674   6.891960   3.951163   6.657840
#> [1831]   3.806635   7.449326   8.597254   8.636239   5.196852   3.816216
#> [1837]   7.485464   2.076168   3.543315   8.008195   6.441080   7.214987
#> [1843]   5.783396   3.558558   8.077792   7.337559   8.349176   6.951736
#> [1849]   6.909446   3.783044   3.867708   4.371238   7.640298   5.665346
#> [1855]   4.253664   4.025155   4.395697   3.041226   7.451819   3.441233
#> [1861]   3.787534   8.112392   6.793874   4.646338   6.468677   8.586126
#> [1867]   4.935040   3.757863   4.742488   7.133235   6.330346   5.822988
#> [1873]   2.555803   6.299370   6.206327   5.408041   3.387633   5.930480
#> [1879]   8.087597   6.319847   3.458757   6.807023   4.834653   8.076926
#> [1885]   6.254258   4.139674   7.153080   3.268945   7.926810   3.127225
#> [1891]   3.133203   6.476832   6.433513   5.916463   5.390871   4.784970
#> [1897]   7.045862   4.784961   4.527763   6.890810   6.009854   8.288781
#> [1903]   4.015584   7.472357   6.454181   4.123732   4.823305   4.465200
#> [1909]   6.376796   3.014555   3.509238   7.876996   3.483332   6.197027
#> [1915]   6.333534   7.225783   3.376553   3.832799   4.448279   6.282072
#> [1921]   7.361032   5.636042   5.482426   2.105748   4.499770   8.180170
#> [1927]   6.674421   2.926363   5.020392   4.539283   3.421941   7.616601
#> [1933]   5.699925   8.503687   3.635933   3.699650   6.557659   5.498680
#> [1939]   7.723716   6.396277   3.680480   6.810858   7.730696   7.370135
#> [1945]   3.870136   4.105451   5.771486   8.668953   3.903307   5.233055
#> [1951]   5.459142   5.219018   4.260367   4.013386   3.562066   7.347813
#> [1957]   5.218653   7.675246   5.824734   6.604248   5.338551   6.743196
#> [1963]   3.962226   5.844991   7.051908   4.649572   4.779830   5.758537
#> [1969]   6.867807   7.440447   4.903929   6.509121   6.614310   5.497818
#> [1975]   7.450721   3.489468   4.836186   4.625556   2.528956   5.880124
#> [1981]   4.456655   6.349494   3.516633   4.478827   4.884793   8.482451
#> [1987]   8.011468   7.464561   5.461835   6.093477   6.784848   5.308867
#> [1993]   3.872026   7.234663   4.464410   7.872009   3.205257   3.164172
#> [1999]   2.768413   4.801355 154.076562  32.882480  39.650246 168.452324
#> [2005]  32.336118 133.505014 116.236703 164.275208  10.302307  23.182107
#> [2011]  18.303494 118.835582  79.511904 149.671324  64.350268 103.672729
#> [2017]  12.511115 147.573640  46.672173  53.839048  26.418543  92.524802
#> [2023] 230.114940  78.821087 106.359676 151.915839  22.873309  42.393949
#> [2029] 223.482574  85.205291  39.881223  77.472228 158.143607  52.164714
#> [2035] 220.705964  61.591464 228.322653  82.984131 140.115244  54.316219
#> [2041]  13.102101  31.052638  64.421223   8.937439 209.671067  31.100199
#> [2047]  39.146534  55.741050  11.360833  35.514097 119.781738 131.873352
#> [2053]  69.775067  48.958797  31.410147  69.903374 203.615864 127.572331
#> [2059]  38.914976 164.181761  76.596551 126.396800 225.448392  21.718713
#> [2065] 149.014855 127.799359 106.083141  13.260995 204.917371  12.819252
#> [2071] 102.781606  28.232552 144.869760  22.285557 170.223541 135.026595
#> [2077]  38.488565  15.959510   9.345202  42.329395  37.912546  23.596385
#> [2083] 149.314952  12.477588 200.096907  55.908873  62.453531 177.598911
#> [2089]  55.473457  98.457499  78.672187  11.478368  90.735475 106.042726
#> [2095]  97.965816  10.356139  64.606468  41.026468  29.562949  31.326077
#> [2101] 212.672571  63.758068  42.982379  11.978354  69.500426  13.235843
#> [2107]  25.640863  91.505031  52.212973 253.248861  12.725523  16.366176
#> [2113]  30.127403  18.936151  12.845181  35.557482  55.309795  54.864215
#> [2119]  82.701441  10.350165 109.008806  60.795060  63.931769 166.154613
#> [2125]  39.148977  77.190633  40.148516  88.564451   8.980168 208.835525
#> [2131] 208.812033  63.633639   9.611855 188.173077  61.417834 148.886613
#> [2137]  12.567574  17.389876 225.480140  44.243052  72.597687 181.207886
#> [2143]  31.991009 153.126909 114.829394  76.032437  72.103371  10.969269
#> [2149]  43.766223  27.625758 232.846291 163.601308  83.945223 202.571325
#> [2155] 113.076150  77.989017  98.877636  16.135282 216.167906 118.410641
#> [2161]  16.675325 147.049063 142.967188  17.574603 100.287484 109.480162
#> [2167] 211.518349 146.297950  62.618993 152.611091 162.425022 310.610995
#> [2173]  51.202174  45.265681 165.805983 237.105954  92.748215 150.434127
#> [2179]  44.775901  60.657683  62.176385  59.175852  45.496223  81.759649
#> [2185] 127.415450 107.326792  62.324835  17.074980  48.522836  79.964352
#> [2191]  17.556749 242.074109 140.682413 170.253307  95.505635 148.977705
#> [2197] 105.299978 253.462590  52.455110  88.842722 116.697816  10.213970
#> [2203]  64.125117  27.601072  25.545121  85.777306  14.630599  76.208696
#> [2209] 112.566942  73.309609  16.108280  18.936025 208.933223  14.820267
#> [2215]  10.062612  12.138097  95.138470 124.890633 191.022151  88.576046
#> [2221]  12.902614  69.112295 135.451453  11.766577  27.014286 118.984894
#> [2227]  33.823334  61.166881  81.881092  30.126628 108.401833  26.900205
#> [2233] 111.308531  61.589119  12.785022 227.752118  18.661311 178.340813
#> [2239]  18.092527  14.223519  55.853563 179.836237 126.252244  81.843300
#> [2245]  84.555010 138.057987  14.008696 140.792981 140.621149 134.557104
#> [2251]  70.143872 156.828862  44.198791  43.948879  12.360672 180.173039
#> [2257]  18.624620  38.755207  42.337596 113.541162  16.115008  82.975719
#> [2263]  15.425706  16.830989 194.377553 215.679175  71.387338 233.364618
#> [2269] 130.956281 129.217216 115.337884  10.598255  25.359491   9.523961
#> [2275] 218.345331  27.483969 148.897155  48.021472  59.019204 134.989699
#> [2281]  24.602201  73.825870  55.164716  16.124882 124.020306  22.680300
#> [2287]  25.517230  89.579122  90.018523 230.872894  54.558407  38.142823
#> [2293]  37.848559  20.437951 187.666285  17.752486 145.184853  90.620335
#> [2299]  69.840921   8.745705  45.584177  29.031389  13.958709 274.491124
#> [2305]  63.317092  22.881456 181.161969  10.960545  35.729473 256.564413
#> [2311]  12.416847 258.305157  11.291199 152.034668  22.074334 103.459260
#> [2317] 269.664648  31.968967  44.345203  23.891353  20.528778 149.138337
#> [2323] 140.471613 200.192921  99.601462 104.846366  11.528106 165.617538
#> [2329]  37.566316  37.611385  13.594388  10.758762 153.335775 143.003002
#> [2335]  92.536130 125.860343  19.954939  13.585757 124.701098 211.025640
#> [2341] 161.200235  33.137915 223.925618  66.335436  49.892296 267.652892
#> [2347]  89.863668 173.894192  46.651793  16.047725  63.282501  29.451052
#> [2353]  82.830963  71.475776 149.369218  23.280427 168.475673  28.796809
#> [2359]  39.647002 152.907289  74.988999  36.172461  49.292779  84.350898
#> [2365]  77.781814  32.885278 109.392459 202.582011  81.548532  18.428257
#> [2371]  60.940227  20.213592 108.423592 109.665258  21.437592  12.552733
#> [2377]  37.865603  35.184271  21.775005   9.846209  16.096827 151.457649
#> [2383] 108.705292  94.530884  45.193439  12.745698 282.571380 153.136703
#> [2389]  10.284517  67.681873  23.721978  57.599029 142.079716 168.511358
#> [2395]  18.024400  84.805838 107.193636  16.968533 148.660636 157.221335
#> [2401] 198.741577 144.816015  48.732654 219.867520  29.745290 141.824815
#> [2407]  71.291070  42.055118 187.225305  59.900378  22.890617  34.978914
#> [2413]  21.755125 190.841922 190.714843 141.385633  93.200203 134.138175
#> [2419] 157.481004  56.791996  10.332770  29.744489 219.319696  69.374491
#> [2425]  29.499319  26.073931 143.276964 108.369647 115.044091 218.688318
#> [2431]  26.416474  52.776237  51.153329  44.862285  66.173158  58.600662
#> [2437]  11.542021 173.491037  70.041381 191.506414  62.605071  80.161892
#> [2443] 187.560217 154.882067  20.169166 124.461475  31.064838  34.238138
#> [2449] 107.795346 224.667615 213.213798  25.551434 218.519586  90.316735
#> [2455]  66.783240  38.755100  12.205615 168.755807  19.104891  81.227670
#> [2461]  10.911247  32.110822  83.282169  74.027890 115.548282  25.278305
#> [2467]  79.876908 108.465255  79.464602 192.582070 207.361562  96.213102
#> [2473] 115.357328  14.094587 123.322697 122.639214 206.856621 148.467519
#> [2479]  66.007289  18.351226 119.729687  87.261560 124.413923  64.570101
#> [2485]  15.771503 122.692997   9.001439  35.201833  15.635592 105.414613
#> [2491] 137.778545 101.478350  29.306799  27.356122  53.064636 137.376067
#> [2497]  51.665064 141.990311  10.127638  17.850085 160.754013  85.875856
#> [2503] 136.956240 120.822628   9.554166 138.754003 209.823752  10.684250
#> [2509]  15.388836 116.197058 208.937893  74.526302  76.302953  54.291139
#> [2515]  72.953247 200.482610  26.648626 124.669142  12.875356   9.280997
#> [2521]  23.674265  92.484952  84.340736 114.490041 108.923777  19.981750
#> [2527]  94.310969 210.973059  16.809861  23.250871 105.968486 230.082387
#> [2533]  22.743568 250.866781 140.933565  15.719761 262.395990 169.208930
#> [2539] 219.912178 142.281542 219.190644  15.580995  14.012132  37.343877
#> [2545]  63.639610  65.016218 189.490667  85.639758  16.661627 138.074116
#> [2551]  91.921263  82.115861  13.878375   9.666904 119.713437 193.835578
#> [2557]  84.295310  12.793018  17.878251 105.840782  12.122056  15.546201
#> [2563]  25.676198  38.858755 116.111038 139.480290  97.259919  52.535683
#> [2569]  29.945014  28.755541  19.912823  24.873872  12.591304 243.801449
#> [2575]  13.055508  70.918166 128.621218  90.879980 115.272101 155.100086
#> [2581] 160.019731 169.992474  14.009220  31.032208  39.125368 170.018630
#> [2587]  97.954287  91.005131  29.765715  32.133761  63.848096  20.142543
#> [2593] 155.449465 171.080929 216.721014 101.362165  49.272809 231.872627
#> [2599]  30.701848 124.475695  52.468914  77.281820  36.320144  18.158675
#> [2605] 192.662189 121.027354  20.528218  16.506379  26.221307 121.498258
#> [2611] 117.749395 125.977086  49.929214  48.986594 202.513999  77.044517
#> [2617] 159.980653  12.532021 263.959195 140.421995  19.383628 238.134975
#> [2623] 170.706356  11.146177  31.837851  36.611048  57.439440  36.705199
#> [2629]  29.240928 163.377387  31.081293  10.339027 216.788527  78.434542
#> [2635]  23.711523  48.702919  17.077268  86.079245 132.111502   9.122157
#> [2641] 216.450639  42.194621 119.688627  73.343673 159.435329 107.378469
#> [2647]  62.462645  12.102365  99.974170  59.188735  79.231262  87.838500
#> [2653]  28.884888  30.083790  13.877581 181.561927  24.184260 225.723021
#> [2659] 223.455217  77.119111  12.928600 113.101009  86.787866 150.089881
#> [2665] 215.235371  60.004139 149.866678 138.433963  13.946052  36.701881
#> [2671]  62.119633  29.518649  28.262600  19.271252 130.301868  24.887602
#> [2677]  71.680353  73.448890  18.986801 162.888107 202.491657 110.090238
#> [2683]  58.334058 183.503905  64.133707  24.098444 101.722262 259.615873
#> [2689]  85.922950  83.167774 122.848071  24.188175 193.057932   8.955880
#> [2695]   9.912721  61.147618  84.161401 119.603108  81.171760 198.952989
#> [2701]  72.139746  26.754117 197.435311  68.975740 201.045883  36.016446
#> [2707] 197.594480 102.656865  31.514803  14.988095 102.764202  59.599638
#> [2713] 212.952083  79.658237  13.596424 146.552474  42.220152  61.651832
#> [2719]  19.880286 243.167110  79.990226  11.472969 195.843554  55.374391
#> [2725] 149.042934 131.850386  20.055065 179.611502 160.638197  26.130656
#> [2731] 116.132845  33.271842   9.656835  41.480844  39.892594  80.044060
#> [2737]  25.283638  80.182027  18.653606 142.794717 210.564723  11.712403
#> [2743] 200.862810  13.380111 204.561663 184.915765  45.442006 197.476321
#> [2749]  63.789058  61.058657  31.844044 164.993260  13.416768  73.456211
#> [2755]  24.134148 132.004558 130.772739   9.208825  15.861593  44.881412
#> [2761]  41.362995  40.727153  32.312274 192.218087  51.263095  85.509356
#> [2767] 120.254971  14.452578   9.398924 174.314440   9.324438  23.021518
#> [2773] 218.186525  67.556407 119.426553  13.593059 215.278900  15.009678
#> [2779]  86.352460  22.182753 180.247046 113.230180  88.487788  62.048716
#> [2785]  33.192107 137.968953 104.781034 215.593341  12.745118  74.144577
#> [2791]  71.545276  21.908615 162.849340  36.214311  10.231953  35.844183
#> [2797] 143.475325  49.160585 108.271582  59.207984 148.935559 162.276507
#> [2803] 147.944919 166.430711  21.947854 136.319447  38.208498 126.234875
#> [2809]  49.354717 168.804040  40.721103  69.599717  51.616707 179.214871
#> [2815] 198.204176 268.601546  14.358410  64.047622  22.437071  96.061234
#> [2821]  96.159479 181.359764 165.431236  68.432824  46.501839 172.702931
#> [2827]   9.955260  90.230161 114.395685  12.442247  48.275699  92.006809
#> [2833]  15.307298  65.510711 171.298615 196.966770  13.200982  88.088799
#> [2839]  69.869936 144.389315 149.517112  71.920993  50.213569 106.019246
#> [2845] 171.390234  93.910101  33.698164 120.036256  55.046983 166.219138
#> [2851]  73.473693 119.450740  72.308059 124.306251  91.569091 208.696546
#> [2857]  29.661005  13.977531  37.386190 194.360561  35.723293 163.193060
#> [2863] 224.141550  60.190071 156.976058  26.913413  55.613759 166.329438
#> [2869]  34.125454 100.495322  94.134603   8.857735  59.384115 199.156697
#> [2875]  10.511118 202.031979  82.676654  44.979149 171.779379  18.494579
#> [2881]  53.444694 220.915609 181.747057  12.784861  42.021752 126.913074
#> [2887]  82.936023  19.983003  33.410016 202.619250  12.335463 171.233511
#> [2893]  53.268115  95.651921 203.651387  40.228198  70.897964 104.012757
#> [2899]  55.002117  78.106082  14.676539  15.129834  13.039554 161.950750
#> [2905] 194.418487  25.150022 113.768943 135.265974  58.944680  95.366171
#> [2911] 131.099950  12.564249  22.425314 219.612973  86.515432 105.167489
#> [2917] 156.328776 240.990235  10.310547  23.801601 194.621995 126.619110
#> [2923]  68.885367 168.705935  10.022582  22.557299  81.018156  29.908764
#> [2929]  56.834667  11.451721  18.100315  58.204790  60.216572  28.130305
#> [2935]  65.779931  78.859015 141.260196  80.993004 209.066326  11.261992
#> [2941]  19.153266 216.175075 193.957378  98.919343 206.533552 122.626066
#> [2947]  60.397945  48.899638 203.691546  61.757859  28.658517  91.589711
#> [2953]  60.051449 108.206461  12.398535  81.349883 212.409742  30.991663
#> [2959] 108.312001 214.831666  13.975444 105.010421 141.209580  82.988349
#> [2965]  78.983136 119.240097 168.138152  50.872174  96.441831  13.775642
#> [2971]  20.034894 191.374075 142.665761 155.453887 142.589786 134.741343
#> [2977]  24.996395 177.743349  16.951348  37.018111  18.132455  78.682608
#> [2983]  39.022428   9.206625  38.909059  51.761874 239.169047  12.775150
#> [2989]  83.274740  25.452958  82.262659  17.229935  69.111805 221.005992
#> [2995]  27.417431 221.747448  29.944942 175.888819 106.958480  70.069846
#> [3001]   9.896206 184.551050  98.734258 158.401109 171.923089 234.568038
#> [3007] 209.455492 221.899075  21.895135  18.292559  34.041480  79.184638
#> [3013] 109.510619  13.790582  12.151920  18.428846  12.939560  82.039600
#> [3019]  24.614037 242.517948 128.250636 131.720873  11.036237 141.071434
#> [3025]  31.522428  13.294601  33.136043  35.265295  72.009880  35.118090
#> [3031]  16.242754  63.089520 397.083329  26.276310  14.589906  10.911882
#> [3037]  35.592697  22.303952 153.962985  96.827947  11.745348  75.479250
#> [3043]  12.897283  80.021180 214.514418  82.903772 136.274373 191.464720
#> [3049]  13.119423 225.751928 131.328164  35.955776  42.460441  25.415450
#> [3055]  12.579469 138.898062 239.385376  90.642704  27.115019 115.751424
#> [3061] 174.594740  56.752818  11.882172 142.247489  63.815769 177.467821
#> [3067] 194.486066   9.002444  26.999597 165.758001 186.171240 195.523355
#> [3073] 107.144313  21.786182 106.280623  21.066092  27.282580  20.798430
#> [3079]  29.445363  57.261802  92.690775 181.330484  22.495040 195.699574
#> [3085]  62.379089  74.732078  79.388815  23.056351  62.000785  21.891100
#> [3091] 112.469312  38.486977  30.767735 135.780182  64.566306  13.177163
#> [3097]  84.427976  81.226545  92.525118  85.389706  53.084051  72.272563
#> [3103]  14.886730  90.779649  20.169081  12.529356  70.287601  14.085123
#> [3109] 158.680641 160.298941  87.444988 203.138804  19.709072  68.576060
#> [3115]  80.766307  59.890891  27.551988 167.706780 233.876229  15.491457
#> [3121]  69.458519 132.502554  68.688728 206.295136 187.403205 137.486635
#> [3127] 139.738856 193.342600  80.244571 201.653620 128.603512  60.816869
#> [3133]  11.121091  23.084186  64.516610  16.416227 126.633953 204.404947
#> [3139] 150.470689  10.645325 125.249506  47.730806  62.957121 169.900046
#> [3145] 176.372877  18.751738  57.983573 124.123964  54.359748  15.803170
#> [3151]  81.625554  12.179816  23.318916  42.386068  57.860632 178.036512
#> [3157]  42.311077  21.120387  52.513517 212.492805  83.302094  15.660159
#> [3163] 204.988878 143.850629  90.844445  14.561585  52.761211 149.954476
#> [3169]  18.491219 101.108604  18.248526  88.200149  76.218710  44.033456
#> [3175]  85.751127  61.183676  66.474431  81.685945 143.240373  64.476061
#> [3181]  81.450597 123.198610 119.545481  84.816262  59.256104  54.495094
#> [3187] 104.836191  13.890432  10.884197 108.086266  72.065163  65.367847
#> [3193] 116.926160  69.026061  13.232624  63.293669 157.719997 108.855912
#> [3199]  43.187221   9.773203  15.753762 117.599410 170.284428  33.491748
#> [3205]  30.971698 184.936764  15.389758 142.792667  51.077064  27.968297
#> [3211]  17.181751 165.122872 102.644654  83.640040  75.535250  15.374440
#> [3217]  28.418696  97.114870  21.789229 165.690862  24.521675  29.180542
#> [3223]  10.914552  32.920721  37.025630 116.460914  86.478184  16.249239
#> [3229] 104.340548 155.358073 103.071976  64.794457  15.110476  10.418744
#> [3235]  95.606533 141.265108 133.879921 231.751294  10.092862  36.301960
#> [3241]   8.686763 265.491846 173.986604  17.383998  21.426732 187.616437
#> [3247] 174.155341  16.991164  11.883685  57.849294  25.574317  24.623387
#> [3253]  73.500435 157.302152  10.019308  35.088228 190.695322  52.497715
#> [3259] 132.529290 202.528885  14.232106 247.254319 212.452643  45.779159
#> [3265]  82.616280  56.117985  11.842699 114.721646  99.737233  27.020373
#> [3271]  37.752973 211.203488  13.284557  51.103854  95.774293  74.140380
#> [3277]   9.343942  93.079010  11.473282  28.283900  35.272375  28.220879
#> [3283]  44.888932  28.291196  63.778381  13.109644 121.480793 141.259411
#> [3289]  46.432613  88.237230 166.173146 114.455433   8.849066  53.078624
#> [3295] 186.969574  44.224414  43.891989  21.023973  65.429660  45.327794
#> [3301] 151.087029 134.890911  89.863936   9.844849 171.138503 156.855095
#> [3307] 178.948555  65.117515  21.538739  22.684720 143.815841 195.306689
#> [3313] 178.422059 143.839053  46.480912  14.222488  97.302070  12.787939
#> [3319]  24.961018  39.454000  10.239395  39.203078  37.397391 120.346680
#> [3325]  10.873483  63.396107 188.868340  21.115767  42.668539  11.286788
#> [3331]  48.433394 103.144669 114.553413 152.045084   9.715978 118.324601
#> [3337] 166.124204  21.315311  21.793013  32.856662  47.217298  35.130930
#> [3343]  15.298184  25.612432 168.924038 149.991150 179.618923  71.029661
#> [3349] 164.962214  31.175181  11.681469  36.458141  44.255607  26.545836
#> [3355]  13.118639  64.445599  56.095291 179.955486 118.305577  16.248109
#> [3361] 204.364571  45.782281 118.908617 228.655255 131.212304  18.563948
#> [3367]  46.890802  19.630686   9.644916 158.122068 134.446353  24.036956
#> [3373]  31.346702 227.026831  74.350833  91.480476  17.050296  53.929647
#> [3379]   9.515604 209.265528  99.536946  20.521860  49.436280 183.579453
#> [3385]  84.415037  34.976476  68.658613 102.340834  38.494028  11.242403
#> [3391]  25.319221  71.412615  71.680130  15.734552 101.897336  24.742484
#> [3397] 195.107046  88.409055  49.851588  62.406411  23.258551  64.157232
#> [3403]  17.639077 113.405893  58.495359  15.766317  52.408462  23.865328
#> [3409]  51.309730  22.630425  49.851450  82.028562  21.172338 159.972029
#> [3415]  25.388371  21.198508  42.180945  19.422063  96.369084  10.381823
#> [3421]  41.200757  17.530007 178.287254  43.106463 231.167314 169.963140
#> [3427] 103.053831  24.922236 109.590395  11.077866 226.815501 234.606003
#> [3433]  96.276364  89.043129  42.612844  92.629902 144.373682  10.566793
#> [3439] 162.154107  88.786994 172.496184  22.112979 244.697152 132.839329
#> [3445]  89.570719  81.320568 124.747497  34.519138  41.664352 209.047176
#> [3451]  12.296675  77.994330   8.933585 142.566109  77.631159  18.504304
#> [3457]  86.715730 149.764656  42.870176  34.276620   9.191538  93.219405
#> [3463] 154.095145 237.069236   9.127463 211.743827  13.374342 143.316867
#> [3469]  12.250839 137.897672 172.994833 188.524981  89.128949 141.453662
#> [3475] 205.354260  32.848726 174.911667  87.336584  53.363641  11.583969
#> [3481]  33.601750  86.019314 148.316350   8.790610  30.226102  54.448609
#> [3487]  10.703862  65.259777 115.708260  46.395747  17.617081  14.199530
#> [3493] 186.530664  44.836520 217.051310 184.048932   8.964830  10.472275
#> [3499]  32.466219  79.021061  10.300509  36.460596  78.681371  12.802134
#> [3505] 153.714548  19.472027  14.480222 160.570644 172.791082  20.731206
#> [3511]  93.093066 228.097985 118.994135 152.658270 221.187204  89.874217
#> [3517]  38.525622  68.866849  23.782867  69.577166 120.288900  50.132891
#> [3523] 138.948773  28.513196  46.509093  22.659924 108.066395 184.583895
#> [3529]  42.973969 192.271371 147.376269   9.433856 130.509221  32.510969
#> [3535]  19.652477  82.349867  24.575423 107.344415  85.364359  27.432038
#> [3541]  22.626717 199.041209  94.451079 222.684066  10.251643 147.997097
#> [3547]  54.423238  49.525198 240.958837 143.978914  79.193510  17.560324
#> [3553] 230.994428 165.290927 130.170749  10.156667  64.687798  54.711025
#> [3559]  43.779640  23.470538 158.552296 166.643064 143.100925  18.432636
#> [3565] 173.244801   9.848271  48.628820 117.236281  10.806656  66.703472
#> [3571]  21.137159  52.231282 129.404078  78.165598  20.333323 227.196252
#> [3577] 129.942131  48.182119 194.594081 138.428489 123.053656  70.750469
#> [3583]  40.775507  69.988476 103.865549  55.975676  47.660173 117.051597
#> [3589]  13.537902 173.428099 100.991659  17.066892  44.355666 204.542011
#> [3595]  30.911943 117.184098 165.583948  82.676227 134.704454 125.491275
#> [3601]  42.509342 203.670995 113.408992  39.698391  30.387709  78.048570
#> [3607]  19.599725 122.671568  34.266985  10.443768  69.617937  17.326636
#> [3613] 106.898667  91.617588 122.846399 127.241443  12.224366  50.287287
#> [3619]  64.246315 143.244856  41.876996  41.405484  24.934713  56.732404
#> [3625] 140.029551 135.088678  90.860778  12.963411  63.286110  76.787309
#> [3631] 122.421669 132.076320  10.151926  16.414995  13.134868 129.241416
#> [3637]  45.569621  49.804194 119.855242 128.937011 175.918684  30.843676
#> [3643] 212.387486  17.388479  47.099927  13.944245 223.287521  74.619716
#> [3649] 105.567165  58.764652 268.364994  64.748935  34.358274 138.241447
#> [3655] 106.385422 159.333539  15.387226  44.754703  78.067204  81.200543
#> [3661]  60.275793 111.712036 171.432306  94.400214  64.899744  51.300372
#> [3667] 162.025869  70.716205 166.772886  31.230254 140.525758 159.775320
#> [3673]  80.731851 105.899230 220.860890  19.987364  45.288195 129.248641
#> [3679]  19.435697 142.185016  21.346975  80.316351   8.986625 104.777508
#> [3685] 108.454913 102.228806 122.606182 184.237167  73.586764 212.891344
#> [3691]  36.003650  14.520914 354.559145  50.547908  50.919652  46.025632
#> [3697]  19.072089  12.157905  18.580033  72.917484  14.046392  41.272836
#> [3703] 202.163827  26.684168  64.683080 113.743682  88.049266 115.568188
#> [3709] 167.703194   8.747358  23.046514  67.826196  47.991697  90.632199
#> [3715]  29.558012  56.382164  97.425665  35.203476 164.070782 202.066311
#> [3721]  82.657289 109.753906  43.936670 104.505039  88.458443  94.227528
#> [3727]  41.564579 141.794439  68.095577 184.723433  61.758085 163.674316
#> [3733] 107.017127  22.203286 216.613060 188.152120  53.675431  40.364353
#> [3739] 128.628387  64.119014  68.926592  85.970755  30.858757 130.029672
#> [3745] 105.159750  68.135469  41.621921 115.318079  22.827520 127.113674
#> [3751]  33.272853  10.371446 108.417926 130.453945  75.379329 102.014197
#> [3757]  89.987257  92.564983 131.241791  28.276948  68.422000  11.379251
#> [3763] 155.595905  93.091877  87.240250 151.262538  68.408730 202.884572
#> [3769] 195.679636 184.841614  23.355197  58.465750 211.481235  16.341084
#> [3775]  71.652806  90.789851 106.288798  13.337351  23.086041   8.704118
#> [3781]  25.738072 203.711584  33.376433 159.401518  59.630699 233.892768
#> [3787]  58.730200  17.951882  23.936560  27.038854  75.950518 178.167212
#> [3793]  90.441093  49.321276  24.582597  74.112282  72.279661  73.098436
#> [3799]  44.421927 231.525666 425.155790  12.136947 157.135567 208.327062
#> [3805] 177.439566  94.321056  37.200056  15.759443  24.134758  33.469832
#> [3811]   9.131537 204.081590  54.643147 102.444612  79.120901  12.764157
#> [3817]  27.498454 108.801896  11.109160  33.220272  13.592632 176.856216
#> [3823] 195.505649  15.391426  26.378780  15.710198  30.372260  19.536252
#> [3829]  84.040417   8.786777  19.375161  27.641780  73.073350 232.195890
#> [3835]  40.848744  28.155230  80.923038  46.146281 176.300879  55.059106
#> [3841]  69.762519 186.877990  28.751929  72.630424 214.184820  27.778799
#> [3847]  92.382359 140.725144  26.242936 184.404053 188.881411  29.649572
#> [3853]  15.663163 141.477039 189.231252  29.439365  40.997488 237.930436
#> [3859]  88.538766  11.720943 171.486850 115.027820  69.298658  68.309748
#> [3865] 161.056805  75.381397  10.547090  35.426891  15.881313  23.712205
#> [3871] 157.639348  23.768165 196.313580  26.895825  81.439888  30.592625
#> [3877] 137.979334  12.454597   9.691855  25.576885   9.941411 132.382689
#> [3883] 170.456694 167.467305 102.859772 172.052602  16.531137  39.130772
#> [3889]  49.442480 207.035420 113.697856 215.967217  37.506796 230.438246
#> [3895] 108.144934 171.618807  91.457290 101.137941  16.743213  63.962056
#> [3901] 210.382877 168.358702  35.091511 173.092313  11.238655  22.267026
#> [3907] 200.309306 133.877908  41.590932 155.424641  21.723192  11.816928
#> [3913]  30.855486  65.787574   9.076546  17.601291  11.717824 187.375511
#> [3919] 216.482718 221.146147  57.040825 244.336555 160.086394  76.026653
#> [3925]  99.735066  67.464582 116.723635  55.851852 169.318729  11.375011
#> [3931] 203.665956  84.286704  44.191156  32.832806  73.233275  44.447527
#> [3937]  81.838800 119.114121  18.642018 200.541884  81.658338  93.164562
#> [3943]  54.706520 109.866474  92.966036  13.695052  62.397617  62.926331
#> [3949]  68.509576 169.127712 144.531212  23.617102   9.035629  45.536757
#> [3955] 116.281133  21.654455  52.566423   8.987314  58.222135 248.482630
#> [3961]  59.998993  49.308740  78.287256  46.197601 221.664083  25.169769
#> [3967]  56.252760 188.313467  48.876401  57.302627 161.869037  30.819444
#> [3973]   8.956110 245.967003 267.760703  73.534316 129.070589  47.063579
#> [3979]  65.704024 229.314138 164.587125  21.172493  29.796419  55.657873
#> [3985] 376.899455  27.668743  31.730308  46.575971 179.429213  48.721687
#> [3991]  16.483578 310.651194  13.300621  23.130836 223.479911 246.401232
#> [3997] 117.300947 164.410788 207.895451  76.241529