Long-term stochastic growth rate worker function for the xsdm model

Computes the negative of the long-term stochastic growth rate, plus log(lambda_max), for the xsdm model, for each location. This is the fast version of a worker function that relies on C++ code, optimized using RcppParallel. The legacy pure-R reference is preserved as xsdm:::like_neg_ltsgr_r for testing and comparison.

Usage

like_neg_ltsgr(
  env_dat,
  mu,
  sigltil,
  sigrtil,
  o_mat,
  num_threads = RcppParallel::defaultNumThreads()
)

Arguments

env_dat

The environmental data array, dimensions (number of locations) x (time series length) x (number of environmental variables). Must not contain missing values.

mu

Vector of optimal environmental values. Length p=dim(env_dat)[3]. Unconstrained real numbers.

sigltil

Vector specifying width of the growth-environment function. Length p=dim(env_dat)[3]. Positive real numbers, Inf entries also allowed.

sigrtil

Vector specifying width of the growth-environment function. Length p=dim(env_dat)[3]. Positive real numbers, Inf entries also allowed.

o_mat

An orthogonal matrix, dimensions p by p.

num_threads

Number of threads for parallel computation. Defaults to RcppParallel::defaultNumThreads().

Value

A vector of length equal to the number of locations, as described above.

Details

Internally, this function:

Reshapes the environmental data into a matrix.

Computes inverse (diagonal) matrices for asymmetry adjustments.

Calls the C++ function like_ltsg() for efficient likelihood computation.

Note

Ensure that env_dat has no missing values. The parameter vectors mu, sigltil, and sigrtil must have length equal to the number of environmental variables (p). The argument o_mat must be a p x p orthogonal matrix (i.e., o_mat %*% t(o_mat) is the identity).

Examples

# Example usage:
like_neg_ltsgr(env_dat = example_1$env_array,
               mu      = example_1$par_list$mu,
               sigltil = example_1$par_list$sigltil,
               sigrtil = example_1$par_list$sigrtil,
               o_mat   = example_1$par_list$o_mat)
#>    [1]  24.582043  78.901260  41.739477  26.530019   3.124899  28.494946
#>    [7]   9.336385   6.902310   6.737008  62.483107  63.465742  63.203659
#>   [13]  10.272974  15.110476   8.579661  61.629050   4.836186  23.067525
#>   [19]   9.785114   7.031786  14.470990  22.521237   8.616595  49.918165
#>   [25]  27.350631  32.388874  65.301127  15.121141  10.266216  33.569972
#>   [31]  37.631710  24.248863  12.747379  17.747061  43.305312  15.062453
#>   [37]  20.913045   5.671054  16.169919   7.464561   4.282575  50.669150
#>   [43]  49.284495   4.090203  11.267579  52.535683  21.034005  43.018176
#>   [49]   5.567738  20.528522  31.266297  14.617560  29.159366   6.735989
#>   [55]   7.929159  24.114338  27.303156  41.506494   9.245740  14.592661
#>   [61]   4.464410  24.843280  60.056210  12.438397   3.557723   3.711949
#>   [67]  47.745296  50.641301  12.201230  11.834848   5.442380  17.955339
#>   [73]  59.304489  47.099532  47.662386  11.956561  30.329664   5.593127
#>   [79]   4.282296  35.039377  31.784194  12.256721  15.719761  12.442808
#>   [85]   4.334486  11.087275  72.205313   4.055621  10.916686  18.196876
#>   [91]  56.469557  30.881256  13.195012  41.529353  76.032437  10.125338
#>   [97]  39.703468   6.054456  52.865072  26.939585  28.458604   7.772086
#>  [103]  28.661609  48.400814   7.735008  15.629499   9.362258   7.723716
#>  [109]  60.658582  25.478477  14.151156  17.594966  10.998391  13.339621
#>  [115]  47.714619   8.704118 103.077588  33.491748  13.189519   9.717430
#>  [121]   4.939287  22.422346  16.361626  14.854086  10.806656   6.302229
#>  [127]  19.325640  13.235843  91.824402   3.693117  64.284303  12.134182
#>  [133]  24.068170  41.794470  37.285184  10.320174  10.890014  14.061843
#>  [139]  63.487971  18.349274  11.316440   5.778746  20.030454   6.142874
#>  [145]  29.180542  67.400516  23.193080  28.429386   5.328191  73.119506
#>  [151]  53.659277  18.112071   5.932482  32.222220  21.636609  10.940924
#>  [157]  18.961096  13.521102   8.489842  22.983697  25.059883  14.805939
#>  [163]   5.889508  54.174978  49.639743  77.447672  61.042630  47.933483
#>  [169]   3.238276  49.929214  59.175715  10.547090  13.765212   7.593603
#>  [175]  57.287059   3.749434   3.162164  34.381343  67.418701  11.978354
#>  [181]  62.189846   5.204241  41.380417   8.972222   4.848766  15.914546
#>  [187]  67.157686   6.921081  39.780329  17.906863  47.774309  11.378865
#>  [193]  26.956975  16.040726  13.412276  19.788906   5.153427  11.589201
#>  [199]   9.772728   6.554315  68.432824  41.590932  12.866871   6.221024
#>  [205]   8.011760  19.013787  42.220152  50.462207  73.559180  13.795813
#>  [211]   8.092929  19.458807  22.764265  12.461052  43.287919  14.548827
#>  [217]  21.358443  28.228715  49.120961   5.518319  50.277659  14.352823
#>  [223]  42.289085  47.835315  72.048094  47.651263  12.832260   9.056501
#>  [229]  24.380104  53.084051   7.308798  15.529461  49.961954  22.304309
#>  [235]   3.912564  15.178879   8.098033  15.368054   3.383786  55.164716
#>  [241]   6.713378   5.309896  39.140742  12.756790  21.900220   5.085363
#>  [247]  47.073313   7.926810  23.493373  20.977446   7.062382  21.665115
#>  [253]   9.676327  19.744228  10.620638  24.273557   4.884793  14.130675
#>  [259]  27.939367  25.452663  40.151686  16.545269  15.519896   5.885004
#>  [265]  11.883499  22.117948   4.308306  56.382164  18.681987  29.468913
#>  [271]  23.705121  18.842932  22.053997   8.440351  56.775410  14.068833
#>  [277]  31.845906   9.611855  18.060928  24.688519  13.571664  25.624610
#>  [283]   6.807023  10.574687  12.787542   8.922026  19.056029   3.684691
#>  [289]  54.316219  14.559383  46.761223  58.713194   3.960390  45.251090
#>  [295]  45.193439  20.946468  17.782682   4.408565  23.603294  14.316840
#>  [301]  58.627091  21.107610  61.396131   7.241586   5.359076   9.280997
#>  [307]   5.328855  17.113698  35.107044  80.021180  15.834835  56.359287
#>  [313]   9.396662  39.130891  19.498207   9.449228  24.766058  42.204675
#>  [319]  13.596424  24.592168  14.242213  32.280876  55.736748  19.658940
#>  [325]   7.472357  19.383628  18.833814   5.758537   7.074372  76.218710
#>  [331]  29.471028  20.142543   8.311488   5.982484  61.952709  13.786862
#>  [337]   7.690488  28.256015  37.611385   7.640298  11.959843  21.023340
#>  [343]  18.630742  19.579577   9.285366  38.739981  40.498425  19.037315
#>  [349]  28.518937  13.748033   9.017299  59.223843 108.807736   7.498210
#>  [355]  19.899947  17.388479   6.910908   6.544578  46.509093   8.077792
#>  [361]  61.166881   5.027541  15.570277  13.305389  11.096010  21.644472
#>  [367]   9.607225  19.153266   6.696599  58.718816  42.285054  85.751127
#>  [373]  44.740002  87.122642  10.635247  15.181652  18.661228  42.833860
#>  [379]  43.377127   7.542512  52.670717  17.850085  36.623032   6.197027
#>  [385]  36.309939   7.832184  22.870324  58.813306   8.507963   5.634191
#>  [391]   7.722776  23.555004   3.611606  21.115767   5.328752  46.799953
#>  [397]   7.330701  18.642628  12.751790   5.410238  10.243650   6.951736
#>  [403]   5.408041   3.237823   3.404963   5.665545  75.435940   5.499401
#>  [409]  49.161133  33.823334  42.386068  62.203715   6.722692  59.224476
#>  [415]   6.944879  10.025733   5.499364   5.496604   6.200861   5.439417
#>  [421]  38.572176  18.838172   5.767246  12.340503  29.445363  22.937026
#>  [427]   9.848908  68.135469   6.279882  28.948875  11.605955  35.055794
#>  [433]  31.068198  33.237311  22.681526  67.741963  31.458788  21.886938
#>  [439]   6.899647  36.663784  27.538759  16.096771   3.600283   3.816216
#>  [445]  39.958162  23.582726  29.784088  31.227707   4.103848  35.658751
#>  [451]  12.319862  21.728368  44.356799   6.355705  22.588146  40.535274
#>  [457]  25.676198  10.099624  14.082727  15.450801   5.385134  10.561305
#>  [463]   7.895934  22.106438  67.955051  51.075264   3.508366  15.302031
#>  [469]  11.564418   3.272479   6.646883  26.771682   9.423296  17.368464
#>  [475]  40.584890  13.378446   9.252358  14.287675  41.760399  12.579639
#>  [481]   2.775246   7.824117   6.290700   6.159364  11.687121   4.892176
#>  [487]  25.415450  15.394698   5.489711  11.207280  50.236159   3.056667
#>  [493]  41.663905  51.041026   7.087695  10.873483  41.914441  57.627156
#>  [499]  16.559033  10.924603  17.817656   9.141870  38.454212  19.224107
#>  [505]  39.234572  12.909497  19.424598  41.442011  12.819055  16.830989
#>  [511]   7.387350  19.962812  12.785022  32.644780  73.456211   6.328251
#>  [517]  20.556225  19.479038   9.005073  54.933136   7.227908  12.923675
#>  [523]   3.862106  35.799518  23.968222  40.965864  35.426891  61.357833
#>  [529]  34.708322  11.442073   9.318678  27.521533  14.431195  60.495197
#>  [535]   9.439925  10.567755  29.619502  31.231314  31.918710  15.046694
#>  [541]  22.260241  11.391400   4.724074  75.353549  24.329069   9.403974
#>  [547]  75.479250  28.020071   8.986625   8.693551   4.308838  13.496000
#>  [553]  14.087607  10.932935   7.735666   6.297229   2.769259  19.945549
#>  [559]  64.908031   4.343616  48.599541  12.764157  23.164642  12.120246
#>  [565]   2.887388  42.194621  47.601251   9.885497  41.190901   6.131482
#>  [571]  59.432762   3.034071   7.448746  11.021673  54.187585   9.896206
#>  [577]  25.059504  52.099207   3.434490  63.969294  52.950099  71.941470
#>  [583]  10.842229  37.605029  15.765437  12.982518  11.360833   3.735960
#>  [589]   7.282657  10.127638  49.804194  23.013246  11.379251   7.639685
#>  [595]  53.261115  14.085123  65.311412   6.251188  33.443323   5.356457
#>  [601]  19.712834  37.441504   7.744252  11.447667  34.836095   4.941145
#>  [607]  33.609231   6.816737  12.663495   7.989933  21.987577   7.273977
#>  [613]  31.393721   9.298714  56.362813   5.430812  53.940594   9.701481
#>  [619]   4.605052   9.713078  22.661544   8.597254   8.519727  29.251824
#>  [625]  10.575580  27.342750   4.682357   5.777626  23.250871   3.994591
#>  [631]   5.540306  21.666315   7.179643  10.576002  24.712535   9.056603
#>  [637]   7.606291  26.239131  19.583128  56.816484  21.285829  23.834089
#>  [643]  18.688800  14.214047   6.151864   9.866080   8.535266  10.168560
#>  [649]  23.662315   9.179256  35.201833  37.114522  17.779203   4.191274
#>  [655]  16.042329  20.597405  18.004861  16.096827   4.028622  23.401812
#>  [661]  37.068886  13.977531  23.437226  15.861593  57.542285  57.481633
#>  [667]  11.722652  82.046395   3.083687  12.951214   9.831036  26.398820
#>  [673]  14.793384  44.355666  66.209654  10.915747   5.729619   4.508214
#>  [679]  11.702291   6.559981   4.973482  53.064636  14.012132   4.736381
#>  [685]  26.734718  68.767816  21.812195  24.646951  11.824670  13.985482
#>  [691]  18.941351   7.003615   6.834511  58.355084  31.777298   6.447468
#>  [697]  22.453701   9.545963   9.912721  13.390211  24.188175  33.912202
#>  [703]  27.814220  27.936392  24.302190  35.327117   6.993114  61.131646
#>  [709]  33.405372   3.851036   8.817083  12.467420  21.199659  44.523576
#>  [715]  27.162024  22.141589  19.090102  10.228491  10.591036  61.866450
#>  [721]   7.463535   8.676462  23.976106   6.861719  22.485554  22.305557
#>  [727]  48.435244   5.170603  50.885697  28.011630  59.256104   8.374321
#>  [733]  11.137707   5.202435   4.746346   8.917430  22.365063  54.662194
#>  [739]  40.164028   6.342421  25.828412   3.217582   7.673952   4.973119
#>  [745]  17.307792  22.494391  20.697064  28.322409   3.951163  37.794856
#>  [751]  19.171241  22.695369  49.321276   8.166651  62.524288  35.413071
#>  [757]   7.912549  32.337260   4.689204  16.726920  15.660159  32.606170
#>  [763]  25.388371  66.703472   8.349176  10.828664   5.077234  55.715101
#>  [769]   4.563356   3.693707  31.081293  10.414104  46.455100   7.825126
#>  [775]   6.342321   4.288629  11.420679  53.556271   7.080314  37.185359
#>  [781]   8.027586   6.947474  50.579023  23.995731  15.389773  19.916044
#>  [787]  13.177163  30.662623  18.837057  15.757298  30.870003  52.513517
#>  [793]  35.723293  39.679007   4.145340  40.913963   8.731456  14.387607
#>  [799]   9.748252  19.378544  55.046983  24.026146  29.908764  67.085949
#>  [805]  35.968470  45.370350   5.794756  11.278113   6.203020  16.047725
#>  [811]  21.254359  21.046786  28.145008  13.887940  47.928910   4.456655
#>  [817]  40.148516  23.719767  28.313692  53.454995  56.360667  13.597228
#>  [823]  19.875123   4.015584  24.742484  11.837088  27.878868  35.635176
#>  [829]  19.412106   4.172477  36.206177  12.120072  25.039771  10.902111
#>  [835]  85.237353  62.177010  54.437000  60.064983  27.838013   5.549308
#>  [841]  30.466713  11.625432  29.649572  51.986053  19.903431  43.891989
#>  [847]  61.085544  28.893828  13.603752  19.076570  15.593902  32.963594
#>  [853]  28.781741   8.138389  60.808070   7.157728   5.161961  49.229167
#>  [859]  72.417855   7.075254  63.958474   4.668416  11.241224  30.514801
#>  [865]  65.890347  40.908901  20.098680   3.477033  15.230112   7.711426
#>  [871]  25.997778  15.515461  11.462353  42.021752   8.251108   8.865976
#>  [877]  37.855734  68.040293   5.200300  44.292420   5.193644  60.704195
#>  [883]   8.596682  37.233900   7.131507  45.037254  31.165371   5.782609
#>  [889]  21.274819   8.580695  27.901661  63.180994  11.104496  32.953932
#>  [895]  79.010497   3.247175   6.589873  58.730200  16.459132  10.284059
#>  [901]   8.081423  15.579816  35.940692  13.596153  12.212936   7.565509
#>  [907]   6.553449  17.540825   5.278206   8.102027   5.207599   4.934277
#>  [913]   7.903943  19.780644  21.445625  36.168251  49.960856  13.740800
#>  [919]  15.913056  50.679730  23.984662  10.884197   7.757152  44.182720
#>  [925]  63.962056  16.351405   3.653575  11.950338  59.233697  15.587945
#>  [931]  55.385636  47.435051  12.137269   8.746070  13.405327   8.318286
#>  [937]   5.233000  32.384883   6.670052  18.018246  23.318916  83.640040
#>  [943]  37.177356  51.263095  34.518748  29.429087  13.504517   6.642817
#>  [949]  55.813923  11.139239   6.823321  18.808528  22.336951  22.518217
#>  [955]   3.212021   7.375325  38.142823  41.121518  28.735198  54.705373
#>  [961]   3.620395   4.707225  34.737777  28.283900  21.513286   8.115789
#>  [967]  35.091511  16.771379  37.334564  21.577053  60.926093  38.176524
#>  [973]  10.127124  34.060198  28.862682  40.689196  33.894819   9.280716
#>  [979]  13.629534  46.603034  41.818666   3.985040  59.673680   5.220890
#>  [985]  13.672906  11.894619  25.574317  29.032711  33.521812  39.041280
#>  [991]   7.759968   3.387633  20.709902  59.630699   5.568095  10.716376
#>  [997]  52.455110   4.108423  52.169220  24.900555  17.013353  73.578927
#> [1003]   3.377473  77.425143  13.655094   5.276290  21.524448 100.051931
#> [1009]  42.289085 223.739572  47.714619 101.140194  13.109460 219.549254
#> [1015]  11.990821  75.357858 123.743451  32.034531 171.779379  20.381709
#> [1021]  77.447125 210.208757  37.414249  99.135738 221.664083 145.371633
#> [1027]  65.429660 141.477039  43.641722 148.398612  15.304090   3.545512
#> [1033]  14.884147 115.924561 147.432855  49.995470  47.701357   8.186124
#> [1039]  81.807115  28.220879  12.464801  18.153278  11.682860  64.464002
#> [1045] 120.096152  31.165371  10.436341  43.186164  19.463092  88.257240
#> [1051]   4.491546 189.031014   4.620082  19.375161 179.492274   8.422796
#> [1057] 168.329702 221.647489  45.926714  51.063629  66.598558 190.734546
#> [1063] 175.787722 200.530752  28.039501 193.205630  62.879324  78.757140
#> [1069]  41.885351 100.350866 235.554725 167.746023  33.837885  52.634893
#> [1075] 233.605651 232.098371 106.068911  86.355928  79.567969 111.042098
#> [1081]  30.858757 122.735002  15.168230  13.954165 136.812459  74.880809
#> [1087]  88.344319  21.636609  22.859943 169.434268  33.912202  41.227518
#> [1093] 218.727289   5.110941  87.168970 246.542204  65.504251 162.838380
#> [1099] 337.335649 107.414722  83.596908  15.628152  12.258116  75.535250
#> [1105] 145.179281  79.618162  54.933136 114.294917 116.644335  18.639397
#> [1111]  20.174794 102.007108  63.286110   2.714523  45.723191  10.272974
#> [1117]   5.338551 224.367827 207.354437 186.653653 158.421910   8.476933
#> [1123]  73.132747  66.173158  19.983003  19.492963  14.798287  20.442412
#> [1129]   4.961691 244.433853   2.329663 149.370002  81.406103  53.205630
#> [1135]   5.220890 189.231527  88.073337  10.489082  47.305341   5.688064
#> [1141] 260.314361 181.876487  11.426323 142.353355  81.236814 143.201271
#> [1147]  92.653106  72.620430 201.963976  52.195384  62.203444  42.975971
#> [1153]  61.596521 223.731539 185.213865  63.246035  57.791489  89.116644
#> [1159]   3.288401   8.362821   3.405037 123.810499   9.993855   7.585584
#> [1165]   2.358384  80.815083 221.248763  41.501848 150.204753  53.717543
#> [1171] 122.938849 166.239704  19.694403  37.114522 155.060914  94.134603
#> [1177]  14.094587  98.094738  86.193753  22.467675  19.435697  13.377230
#> [1183] 101.981692 172.664445  17.471175 237.501733  47.458518 173.428099
#> [1189]  24.582597  54.843351 182.112164   7.922041  61.131646   9.942126
#> [1195] 244.799629  24.036956 132.192679  52.886939 153.838312  71.263865
#> [1201]  14.739196  20.102632 219.283241  96.260260  49.308740  89.811073
#> [1207]  74.770578 103.210350 167.373129 121.926429   6.330346  76.183985
#> [1213]  76.327599  16.475393  13.349220   3.169437  64.532843  71.583571
#> [1219] 234.141474  50.576715   3.684691  51.075264 214.578201 104.345608
#> [1225] 173.424575  35.557482  59.019204  25.292514 186.156902  64.441452
#> [1231]  10.756010 175.716812 175.002791  23.628058 157.885293  93.060929
#> [1237]  93.764679  46.588416  64.228703  13.953224  74.146557  77.492187
#> [1243] 204.878907  89.682289  36.680932 118.514356  18.774649 208.618513
#> [1249] 112.469312   7.799905  40.775507 215.278900  66.662194   5.192749
#> [1255] 124.331629   7.607018  49.133208  76.241529 202.799850  10.356139
#> [1261]  26.423341  87.020019 245.160150  10.250975 151.410213   8.600092
#> [1267] 226.956200 193.200849  12.513225 270.799596 260.975567  58.535505
#> [1273] 171.009113  23.397642  90.840106 196.249118  23.463328 259.615873
#> [1279]  43.305312 104.846366 258.954054 158.819804  32.920721  24.025438
#> [1285] 116.193836  19.658940  12.928600  48.671008  41.595339 167.645889
#> [1291]  26.061018   7.498210 110.335286 167.133010  76.054839  12.974009
#> [1297] 159.163179  14.323092   8.846875  18.838172  53.940594  81.132946
#> [1303] 118.264243   4.802776  95.219853  84.376424  11.233480 143.003002
#> [1309] 192.624288 206.907090  69.577166   8.416415  28.518937  84.180360
#> [1315] 193.550085   6.876833  37.642307  78.308444  43.239417  49.918165
#> [1321]   9.669754   9.302804  38.469754   6.167402  94.079223  21.178950
#> [1327]  14.392703  10.231953  16.144089  73.034266 208.955911 193.906577
#> [1333] 129.448746 117.672012  81.439888 226.095159  29.735393  15.342698
#> [1339]  82.045098  16.474389  52.454575 108.044417 215.679175  51.832355
#> [1345]   3.278781  44.076640   7.258635  90.655077  24.351081   7.639685
#> [1351]  28.895687  41.380417 128.937011   6.083392 128.346077 220.061021
#> [1357]  15.388836  36.319264  59.364510  69.268734 116.804649  10.239395
#> [1363] 106.323670 145.584930 281.366174  65.681343  62.414972 180.748206
#> [1369]   4.142912   4.328956   6.202068   7.893728  11.993609  34.836095
#> [1375]  96.369084   3.971800 134.791856 225.814460  63.465742  52.052292
#> [1381]   3.117300 268.601546 190.194819   4.224112   5.931950  19.903431
#> [1387] 104.214138 137.944252  19.259233 188.493707 190.219721  63.118329
#> [1393]   6.733263  16.808209   9.668310  11.834877  15.100165 119.865971
#> [1399]  26.530019  89.753304 169.118861  49.475021  99.371949 341.425193
#> [1405] 119.345721   6.664898   4.620366 228.380335 182.118033 175.918684
#> [1411]   8.667614  39.031432 206.336340   9.833755   8.081423  12.868818
#> [1417] 201.836019  80.619283  29.715555   6.407779 105.407898  23.055052
#> [1423]  25.920463  84.040417  16.661627 177.717154  15.861593 179.836237
#> [1429] 101.970496  41.818666  17.434350 218.959229 126.571778 232.869063
#> [1435]  38.732092   4.259917 130.345978  27.919008  54.503819  28.568085
#> [1441]  20.145240  31.673656  40.800589  53.074456  70.111150   5.394256
#> [1447]  69.634957 106.898667   6.993322   3.345116  90.018523 108.529376
#> [1453]  51.409548 112.718140 109.055767 126.442689  44.043586  27.551988
#> [1459]  52.495512  17.744250 153.854178  28.027758 116.483980   8.536342
#> [1465] 170.706356   7.120062 104.203232  69.556483  62.092528  19.579577
#> [1471]   3.845642  16.510406  29.453085 223.612989  52.361730  12.477588
#> [1477] 160.831360  11.882172 204.521851 165.388366 138.746849 202.138867
#> [1483] 134.446353  53.925181 173.986604  35.580047 221.407258  67.449867
#> [1489]   4.651365 231.791018  13.786956  24.931353 141.372619 273.237889
#> [1495]  29.251824  93.204357   7.126731  58.982963   3.737604   8.099398
#> [1501] 113.675586 228.858579 106.288798   4.033387   8.076926  40.346861
#> [1507] 221.635899 147.149672 184.820747 149.894750   7.964937 153.383115
#> [1513] 118.460587  85.658563  15.169590  67.158948  16.248109  65.779931
#> [1519]  84.544420 186.921066  19.277110 104.536045  13.037415 194.503493
#> [1525]  42.881672  41.323490  10.003825   3.407620  69.019100  72.925533
#> [1531] 211.433225  84.360993 208.219786 105.558597  32.346630  98.011780
#> [1537] 206.295851 106.707896  19.090102   2.965375  15.668325  44.272703
#> [1543]  23.534653 108.855912 192.469480  25.886154  21.483904  16.978201
#> [1549] 116.788377  19.411193  36.011863   3.948832   3.377781   3.250422
#> [1555] 214.615968 230.607111  42.386068 198.661634 174.024261  98.276777
#> [1561]  17.178734  83.870572  49.048212  20.194395   5.636042 215.756374
#> [1567] 185.665703  86.141340  18.751738   6.103555 178.340813   9.989085
#> [1573]   6.998908 242.809138  29.945014  53.363641  33.894819 123.130607
#> [1579]  76.683278  40.148516 117.122294 168.677896  84.936835  63.616675
#> [1585]  23.547492 151.262538 112.255481 221.577779 210.141584   2.743534
#> [1591] 127.107721  21.508921   6.997027  12.374303  86.809930  40.620653
#> [1597]  22.084062 125.977086  11.495942  27.356122  47.413561  50.393176
#> [1603]   5.169769 116.051448  13.900487  59.946439  88.957944  29.429087
#> [1609] 145.898609  48.435244 210.252897  18.060928   2.997575  79.990226
#> [1615]  67.773479 175.041977   4.983289 156.328325   9.266557   6.494429
#> [1621]  68.576060   9.439925  27.680998  54.419719 118.981174   2.418845
#> [1627]  10.207365 171.486850  29.765715   5.548968  20.127230 137.979334
#> [1633]  55.797289 104.916703  24.299351  59.246125  63.866847   3.384265
#> [1639]  78.597033  88.304939  12.251728  23.936560  25.980659  14.352823
#> [1645]  29.092129 173.636424  23.525289   3.517795 180.292086  53.617565
#> [1651]  92.045741  66.884710 122.766089  15.174267  11.529343  94.451079
#> [1657]  75.056099  92.690775  22.465230   4.699454   3.068674   9.133721
#> [1663]   4.787552  87.454494   5.751400  45.922976   3.084476 102.461843
#> [1669]  73.743741  11.687121  23.538029  15.090191  70.280719  23.006700
#> [1675]  18.638882  24.185720 110.807318  64.410584  75.854395  29.340487
#> [1681] 141.257272 210.926989  18.454091  12.441453 181.176082  97.887062
#> [1687] 130.264102 209.143049  15.083601  14.431195  40.493870 215.967217
#> [1693] 101.362165 122.045727  75.725928   3.341317 272.573930  27.641780
#> [1699]  86.559155  14.027068   3.558558  13.080894   9.329521  24.646951
#> [1705]   4.025248 182.074850  11.894619  39.125368   6.388729 140.261799
#> [1711]  12.529356   8.845372  11.469156   7.735666   5.362398 116.697816
#> [1717]  51.855030 128.186917 168.272148   8.903294  21.111405 111.431792
#> [1723]   8.022582 208.123292  87.164485  54.864949  40.228198  30.162387
#> [1729]   3.434915   5.497818  91.967669  96.806438 211.203488 198.003941
#> [1735] 113.405893   6.412783  30.911943 101.112718  41.590932  81.856925
#> [1741]  59.333301  37.126837 297.259071 178.245193  10.805513  55.988619
#> [1747]  72.342620  12.138097  30.466713  18.316189 142.282960  21.603690
#> [1753] 118.305577  61.397824   7.740601   3.897595   5.312708  15.015542
#> [1759] 142.185016  52.090489  26.653583 123.807917 228.827420  38.722604
#> [1765]  48.481115  10.251643  18.695534   2.794768  43.779640  53.758460
#> [1771] 191.506414 170.488115  16.351405 242.484445 113.697856  64.670850
#> [1777]  15.209703 108.063721 181.359764  69.374491  19.378544  33.469832
#> [1783]  39.703468  42.345031  45.425926 237.349092 123.903499 136.705706
#> [1789] 254.223448  52.670717 124.413923  24.399691 100.513224  13.012420
#> [1795]  41.564579  10.306193   6.299538 270.831853  37.801472  24.309320
#> [1801]  11.898798   7.511881 202.491657  32.280876   7.094541  64.092430
#> [1807]   5.498680  46.620120  26.946472 206.875583  62.462645 160.430090
#> [1813]  15.881313 188.313467  32.336118  61.297280  51.103249   5.619550
#> [1819] 124.764314 244.697152  21.482748   7.864908 143.494085  15.680030
#> [1825]   6.640334   9.021711 169.844362 133.332832 223.040868 172.004421
#> [1831]  14.234910 152.556100 103.233671 201.170503 155.427993  55.042112
#> [1837]  22.659875 212.569451   5.170750  87.490348  29.487959  23.005359
#> [1843]  34.918997  45.294043  90.441093  87.443396 256.182999  16.743213
#> [1849]  11.045491 162.005397  22.819356 251.985522  23.659386  18.180142
#> [1855]   6.142874  38.016929 161.260950 122.463376 139.362794  17.628499
#> [1861]  35.088228   6.649858   5.916463  65.953476 121.650994 209.281501
#> [1867]  12.630273  30.042300  67.446581 213.502504  43.000272   9.916279
#> [1873]  66.209654   2.520724 189.115295  83.302094 100.816246  55.473457
#> [1879]   5.848321  71.757874 161.722272  20.440565  46.032200  22.630512
#> [1885]  37.098125  78.518547 108.361499   8.752290   9.727627 154.936091
#> [1891] 197.644194  10.537413  26.242936  32.322068  22.095308  13.694738
#> [1897]  96.190900  11.447928  16.459132   5.443358  30.436881  70.352772
#> [1903]  24.540543  98.327843  15.979558   8.163568  28.751929 197.312054
#> [1909] 189.114157  56.222111  11.295646 274.504492   4.302055 268.570130
#> [1915]  19.458805  61.013373 175.269796  16.375768  50.082927  13.387980
#> [1921]  69.496829  75.476337  18.349274  17.400604   6.749103  14.805939
#> [1927]  21.047230  71.312345 205.834289  18.820569   3.674738 128.789848
#> [1933]  66.128197 124.770347   9.810882 196.332276  22.681526 147.983960
#> [1939] 192.944598  12.224366 219.460721  86.031183  22.275799  24.109868
#> [1945]  96.471970 195.755144  11.656030  58.219572  61.589119   8.469149
#> [1951]  70.478088  31.029185  46.444042   6.141461 124.975302  35.018656
#> [1957] 160.086394  31.439500 181.552050 184.577708  17.782682  44.198791
#> [1963]  86.013575  63.385543   9.443914 119.579225  44.356799   5.978200
#> [1969]   8.134278 227.284661  28.023462  45.635860  21.383332  37.968365
#> [1975]  40.746259   3.126129 166.772886 147.575054  77.301648  24.934713
#> [1981]  10.673471  21.780175   5.185237 110.014823 182.679358   3.712219
#> [1987] 102.075925  22.536872   4.073289   4.921445  76.909213 129.441410
#> [1993]  10.332770 203.615864  33.354038 152.884387  20.522933  18.009839
#> [1999]   3.470315 124.238046