xsdm: A Demographic Approach to Species Distribution Model

Integrates concepts of stochastic demography into species distribution modelling. Inferences, utilizing Bayesian methods based in 'Stan', are made of aspects of the environmental niche of a species, and those inferences can be used to project species geographic range. Data requirements include species detection records and spatiotemporal environmental data in Berti et al. (2024, doi:10.1101/2024.10.30.621023 ) for a description of the statistical and demographic models used.

See also

Author

Maintainer: Angel Luis Robles Fernandez a.l.robles.fernandez@gmail.com (ORCID)

Authors:

Emilio Berti emilio.berti@idiv.de (ORCID)

Daniel C. Reuman reuman@ku.edu (ORCID)